% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{completevalues}
\alias{completevalues}
\title{Complete values in groups without numeric information}
\usage{
completevalues(d, var_num)
}
\description{
Complete values in groups without numeric information
}
