% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_draw}
\alias{data_draw}
\title{Create data for a Highcharts plot}
\usage{
data_draw(
  data,
  dic = NULL,
  var_cat = NULL,
  var_num = NULL,
  var_date = NULL,
  viz = NULL,
  frType = NULL,
  opts = NULL
)
}
\arguments{
\item{data}{a data frame containing the data to be plotted}

\item{dic}{a named vector specifying the categories and colors for the data}

\item{var_cat}{the name of the categorical variable in the data frame}

\item{var_num}{the name of the numeric variable in the data frame}

\item{var_date}{the name of the date variable in the data frame}

\item{viz}{the type of visualization to be created ("treemap", "line", "bar", "pie", or "donut")}

\item{frType}{a string indicating the format of the data frame (e.g., "CatNum", "DatNum", "CatCatNum")}

\item{opts}{a list containing options for the plot (e.g., tooltip_template, na_label, format_sample_num)}
}
\value{
a list of data for use in creating a Highcharts plot
}
\description{
This function creates data for a Highcharts plot, given input data and options.
}
\keyword{internal}
