% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\docType{data}
\name{gg_test_docs}
\alias{gg_test_docs}
\title{gg_test_docs
test}
\format{An object of class \code{NULL} of length 0.}
\usage{
gg_test_docs
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
gg_test_docs
test
}
\section{noftypes}{
 Ca,Ca-Nu
}
\examples{
add(1, 1)
}
\keyword{datasets}

