% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_2yline_YeaNumNum}
\alias{hgch_2yline_YeaNumNum}
\title{2 y line}
\usage{
hgch_2yline_YeaNumNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle1 = NULL, yAxisTitle2 = NULL,
  symbol = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
2 y line
}
\section{ftypes}{
 Yea-Num-Num
}

\examples{
hgch_2yline_YeaNumNum(sampleDatta("Yea-Num-Num",nrow = 10))
}
