% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_2yline_YeNuNu}
\alias{hgch_2yline_YeNuNu}
\title{hgch_2yline_YeNuNu
2 y line}
\usage{
hgch_2yline_YeNuNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle1 = NULL, yAxisTitle2 = NULL,
  symbol = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_2yline_YeNuNu
2 y line
}
\section{ftypes}{
 Ye-Nu-Nu
}
\examples{
hgch_2yline_YeNuNu(sampleData("Ye-Nu-Nu",nrow = 10))
}

