% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_Cat}
\alias{hgch_area_Cat}
\title{Area (categories)}
\usage{
hgch_area_Cat(data = NULL, agg_text = NULL, allow_point = FALSE,
  background = "#ffffff", border_color = "#CCCCCC", border_width = 1,
  caption = NULL, clickFunction = NULL, colors = NULL,
  color_click = NULL, color_hover = NULL, color_opacity = 0.7,
  cursor = NULL, drop_na = FALSE, export = FALSE,
  fill_opacity = 0.5, highlight_value = NULL,
  highlight_valueColor = "#F9B233", horLabel = NULL, horLine = NULL,
  horLine_label = " ", labelWrap = 12, lang = "es", marks = c(".",
  ","), nDigits = NULL, null_color = "#f7f7f7", order = NULL,
  orientation = "ver", percentage = FALSE, prefix = NULL,
  text_show = TRUE, sliceN = NULL, sort = "no", spline = FALSE,
  startAtZero = TRUE, subtitle = NULL, suffix = NULL, title = NULL,
  theme = NULL, tooltip = list(headerFormat = NULL, pointFormat =
  NULL), verLabel = NULL, verLine = NULL, verLine_label = " ",
  opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_area_Cat(sampleData("Cat", nrow = 10))
}
