% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_CatCat}
\alias{hgch_area_CatCat}
\title{Area (categories, ordered categories, numbers)}
\usage{
hgch_area_CatCat(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat
}

\examples{
hgch_area_CatCat(sampleData("Cat-Cat-Num", nrow = 10))
}
