% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_CatDatNum.R
\name{hgch_area_CatDat}
\alias{hgch_area_CatDat}
\title{area Chart Cat Dat}
\usage{
hgch_area_CatDat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
area Chart Cat Dat
}
\section{ctypes}{

Cat-Dat
}

\examples{
hgch_area_CatDat(sampleData("Cat-Dat", nrow = 10))
}
