% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_CatDatNum}
\alias{hgch_area_CatDatNum}
\title{Vertical area (categories, dates, numbers)}
\usage{
hgch_area_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among categories over dates
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
hgch_area_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
