% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_CatYeaNum.R
\name{hgch_area_CatYea}
\alias{hgch_area_CatYea}
\title{area Chart Cat Yea}
\usage{
hgch_area_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
area Chart Cat Yea
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_area_CatYea(sample_data("Cat-Yea", nrow = 10))
}
