% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_DaNu}
\alias{hgch_area_DaNu}
\title{hgch_area_DaNu}
\usage{
hgch_area_DaNu(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  symbol = NULL, startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_area_DaNu
}
\section{ftype}{
 Da-Nu
}
\examples{
hgch_area_DaNu(sampleData("Ca-Da-Nu",nrow = 10))
}

