% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_Dat.R
\name{hgch_area_Dat}
\alias{hgch_area_Dat}
\title{Area chart Dat}
\usage{
hgch_area_Dat(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read \code{\link[dsvizopts]{chart_viz_options}} a general options summary to configure your hgchmagic plots.}
}
\description{
\code{hgch_area_Dat()} Create a highcharter area plot based on a particular data type.
In this case, you can load data with only one \strong{date column} or be sure of
the first column of the dataframe its categorical since it will be done a
counting the categories of this column
}
\section{Ftype}{

Dat
}

\examples{
data <- sample_data("Dat", n = 30)
hgch_area_Dat(data)

# data with more of one column
data <- sample_data("Dat-Num-Dat-Dat-Dat", n = 30)
hgch_area_Dat(data)

# calculate percentage
hgch_area_Dat(data, percentage = TRUE)

# You can call the count and percentage in the tooltip plot
data \%>\%
hgch_area_Dat(tooltip = "Count: {Count} <br/> Percentage: {\%}\%")

}
\seealso{
Other Dat plots: 
\code{\link{hgch_line_Dat}()},
\code{\link{hgch_scatter_Dat}()}
}
\concept{Dat plots}
