% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_DatNum}
\alias{hgch_area_DatNum}
\title{Vertical area}
\usage{
hgch_area_DatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical area
}
\section{ftype}{

Dat-Num
}

\examples{

hgch_area_DatNum(sampleDatta("Cat-Dat-Num",nrow = 10))

}
