% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_OcaNum}
\alias{hgch_area_OcaNum}
\title{Vertical area (ordered categories, numbers)}
\usage{
hgch_area_OcaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  agg = "sum", dropNa = FALSE, order = NULL, percentage = FALSE,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among categories
}
\section{ctypes}{

Oca-Num
}

\examples{
hgch_area_OcaNum(sampleData("Cat-Num", nrow = 10))
}
