% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_OcaNumP}
\alias{hgch_area_OcaNumP}
\title{Area (ordered category, n numbers)}
\usage{
hgch_area_OcaNumP(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  agg = "sum", dropNa = FALSE, order = NULL, percentage = FALSE,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Oca-NumP
}

\examples{
hgch_area_OcaNumP(sampleData("Oca-NumP", nrow = 10))
}
