% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stack_100_CaDaNu}
\alias{hgch_area_stack_100_CaDaNu}
\title{hgch_area_stack_100_CaDaNu}
\usage{
hgch_area_stack_100_CaDaNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, symbol = NULL, startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_area_stack_100_CaDaNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_area_stack_100_CaDaNu(sampleData("Ca-Da-Nu",nrow = 10))
}

