% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stack_CaCaNu}
\alias{hgch_area_stack_CaCaNu}
\title{hgch_area_stack_CaCaNu}
\usage{
hgch_area_stack_CaCaNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, symbol = NULL, startAtZero = FALSE, export = FALSE,
  ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_area_stack_CaCaNu
}
\section{ftype}{
 Ca-Ca-Nu
}
\examples{
hgch_area_stack_CaCaNu(sampleData("Ca-Ca-Nu",nrow = 10))
}

