% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stack_CaYeNu}
\alias{hgch_area_stack_CaYeNu}
\title{hgch_area_stack_CaYeNu}
\usage{
hgch_area_stack_CaYeNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, symbol = NULL, startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_area_stack_CaYeNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_area_stack_CaYeNu(sampleData("Ca-Ye-Nu",nrow = 10))
}

