% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stacked_100_CaDaNu}
\alias{hgch_area_stacked_100_CaDaNu}
\title{Vertical 100% stacked area}
\usage{
hgch_area_stacked_100_CaDaNu(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical 100% stacked area
}
\section{ftype}{
 Ca-Da-Nu
}

\examples{
hgch_area_stacked_100_CaDaNu(sampleData("Ca-Da-Nu",nrow = 10))
}
