% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stacked_100_CatCatNum}
\alias{hgch_area_stacked_100_CatCatNum}
\title{Vertical 100 stacked area}
\usage{
hgch_area_stacked_100_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical 100 stacked area
}
\section{ftype}{
 Cat-Cat-Num
}
\examples{
hgch_area_stacked_100_CatCatNum(sampleDatta("Cat-Cat-Num",nrow = 10))
}

