% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stacked_CatDatNum}
\alias{hgch_area_stacked_CatDatNum}
\title{Vertical stacked area}
\usage{
hgch_area_stacked_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical stacked area
}
\section{ftype}{
 Cat-Dat-Num
}

\examples{
hgch_area_stacked_CatDatNum(sampleDatta("Cat-Dat-Num",nrow = 10))
}
