% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stacked_CatOcaNum}
\alias{hgch_area_stacked_CatOcaNum}
\title{Vertical stacked area (categories, ordered categories, numbers)}
\usage{
hgch_area_stacked_CatOcaNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among stacked categories
}
\section{ctypes}{

Cat-Oca-Num
}

\examples{
hgch_area_stacked_CatOcaNum(sampleData("Cat-Cat-Num", nrow = 10))
}
