% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_CatYeaNum.R
\name{hgch_area_YeaCatNum}
\alias{hgch_area_YeaCatNum}
\title{area Chart Yea Cat Num}
\usage{
hgch_area_YeaCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
area Chart Yea Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_area_YeaCatNum(sample_data("Yea-Cat-Num", nrow = 10))
}
