% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_YeNu}
\alias{hgch_area_YeNu}
\title{hgch_area_YeNu}
\usage{
hgch_area_YeNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_area_YeNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_area_YeNu(sampleData("Ye-Nu",nrow = 10))
}

