% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_Cat}
\alias{hgch_bar_Cat}
\title{Bar (categories)}
\usage{
hgch_bar_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  colors = NULL, colorText = "#0E0329", colorScale = "no",
  dropNa = FALSE, tooltip = list(headerFormat = NULL, pointFormat =
  NULL), format = c("", ""), highlightValue = NULL,
  highlightValueColor = "#F9B233", labelWrap = 12, marks = c(".",
  ","), nDigits = NULL, order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, export = FALSE,
  theme = tma(), ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_bar_Cat(sampleData("Cat", nrow = 10))
}
