% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatCat}
\alias{hgch_bar_CatCat}
\title{Grouped bar by first category (categories, categories)}
\usage{
hgch_bar_CatCat(
  data,
  allow_point = FALSE,
  background = "#ffffff",
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_opacity = 0.7,
  color_scale = "discrete",
  cursor = NULL,
  drop_na_v = c(FALSE, FALSE),
  export = FALSE,
  graph_type = "grouped",
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = " ",
  label_wrap_v = c(12, 12),
  lang = "es",
  legend_position = "center",
  legend_show = TRUE,
  marks = c(".", ","),
  n_digits = NULL,
  null_color = "#f7f7f7",
  order1 = NULL,
  order2 = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bar_CatCat(sampleData("Cat-Cat", nrow = 10))
}
