% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatNum.R
\name{hgch_bar_CatNum}
\alias{hgch_bar_CatNum}
\title{Bar Chart Cat Numeric}
\usage{
hgch_bar_CatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Num, Yea-Num
}

\examples{
hgch_bar_CatNum(sample_data("Cat-Num", nrow = 10))
}
