% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatNumP}
\alias{hgch_bar_CatNumP}
\title{Bar (ordered category, n numbers)}
\usage{
hgch_bar_CatNumP(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Cat-NumP
}

\examples{
hgch_bar_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
