% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatCat.R
\name{hgch_bar_CatYea}
\alias{hgch_bar_CatYea}
\title{bar Chart Cat Yea}
\usage{
hgch_bar_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bar Chart Cat Yea
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bar_CatYea(sample_data("Cat-Yea", nrow = 10))
}
