% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatCatNum.R
\name{hgch_bar_CatYeaNum}
\alias{hgch_bar_CatYeaNum}
\title{bar Chart Cat Yea Num}
\usage{
hgch_bar_CatYeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bar Chart Cat Yea Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bar_CatYeaNum(sample_data("Cat-Yea-Num", nrow = 10))
}
