% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_DatNum.R
\name{hgch_bar_DatNum}
\alias{hgch_bar_DatNum}
\title{Bar Chart Dat Numeric}
\usage{
hgch_bar_DatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Dat-Num
}

\examples{
hgch_bar_DatNum(sampleData("Dat-Num", nrow = 10))
}
