% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_CatCat}
\alias{hgch_bar_grouped_CatCat}
\title{Grouped bar by first category (categories, categories)}
\usage{
hgch_bar_grouped_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = NULL, verLine = NULL, verLineLabel = NULL,
  dropNa = c(FALSE, FALSE), order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", format = c("", ""),
  sliceN = NULL, theme = tma(), tooltip = list(headerFormat = NULL,
  pointFormat = NULL, shared = NULL), export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bar_grouped_CatCat(sampleData("Cat-Cat", nrow = 10))
}
