% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_CatCatNum}
\alias{hgch_bar_grouped_CatCatNum}
\title{Grouped bar by first category (categories, ordered categories, numbers)}
\usage{
hgch_bar_grouped_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = NULL, verLine = NULL, verLineLabel = NULL,
  agg = "sum", nDigits = NULL, marks = c("", "."),
  dropNa = c(FALSE, FALSE), legendLayout = "horizontal",
  order = NULL, orientation = "ver", percentage = FALSE,
  sort = "no", format = c("", ""), sliceN = NULL, theme = tma(),
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
hgch_bar_grouped_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
