% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_CatYeaNum}
\alias{hgch_bar_grouped_CatYeaNum}
\title{Grouped bar (categories, year, numbers)}
\usage{
hgch_bar_grouped_CatYeaNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = NULL, verLine = NULL, verLineLabel = NULL,
  agg = "sum", nDigits = NULL, marks = c("", "."),
  dropNa = c(FALSE, FALSE), legendLayout = "horizontal",
  order = NULL, orientation = "ver", percentage = FALSE,
  sort = "no", format = c("", ""), sliceN = NULL, theme = tma(),
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among categories over years
}
\section{ctypes}{

Cat-Yea-Num
}

\examples{
hgch_bar_grouped_CatYeaNum(sampleData("Cat-Yea-Num", nrow = 10))
}
