% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_hor_CaNuP}
\alias{hgch_bar_grouped_hor_CaNuP}
\title{Horizontal grouped bar}
\usage{
hgch_bar_grouped_hor_CaNuP(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal grouped bar
}
\section{ftypes}{
 Ca-NuP
}

\examples{
hgch_bar_grouped_hor_CaNuP(sampleData("Ca-NuP",nrow = 10))
}
