% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_hor_CatNum}
\alias{hgch_bar_grouped_hor_CatNum}
\title{Horizontal grouped bar}
\usage{
hgch_bar_grouped_hor_CatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  aggregation = "sum", startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal grouped bar
}
\section{ctypes}{

Cat-Num
}

\examples{
hgch_bar_grouped_hor_CatNum(sampleData("Cat-Num", nrow = 10))
}
