% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_ver_CatYeaNum}
\alias{hgch_bar_grouped_ver_CatYeaNum}
\title{Vertical grouped bar}
\usage{
hgch_bar_grouped_ver_CatYeaNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical grouped bar
}
\section{ftype}{
 Cat-Yea-Num
}
\examples{
hgch_bar_grouped_ver_CatYeaNum(sampleDatta("Yea-Num-Num",nrow = 10))
}

