% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_ver_CaYeNu}
\alias{hgch_bar_grouped_ver_CaYeNu}
\title{hgch_bar_grouped_ver_CaYeNu}
\usage{
hgch_bar_grouped_ver_CaYeNu(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_grouped_ver_CaYeNu
}
\section{ftype}{
 Ca-Ye-Nu
}

\examples{
hgch_bar_grouped_ver_CaYeNu(sampleData("Ye-Nu-Nu",nrow = 10))
}
