% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_bar_hor_c}
\alias{hgch_bar_hor_c}
\title{hgch_bar_hor_c}
\usage{
hgch_bar_hor_c(f, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  sort = "no", aggregate = "count", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_hor_c
}
\section{ftype}{
 Ca
}
\examples{
hgch_bar_hor_c(sampleData("Ca",nrow = 10))
}

