% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_hor_Cat}
\alias{hgch_bar_hor_Cat}
\title{Horizontal bar}
\usage{
hgch_bar_hor_Cat(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = "count", theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal bar
}
\section{ctypes}{

Cat
}

\examples{
hgch_bar_hor_Cat(sampleDatta("Cat",nrow = 10))
}
