% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_bar_hor_top_c}
\alias{hgch_bar_hor_top_c}
\title{hgch_bar_hor_top_c}
\usage{
hgch_bar_hor_top_c(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, reverse = TRUE, aggregate = "count", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_hor_top_c
}
\section{ftype}{
 Ca
}
\examples{
hgch_bar_hor_top_cn(sampleData("Ca-Nu",nrow = 10))
}

