% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_hor_top_Ca}
\alias{hgch_bar_hor_top_Ca}
\title{Ordered horizontal bar}
\usage{
hgch_bar_hor_top_Ca(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, reverse = TRUE,
  aggregate = "count", theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Ordered horizontal bar
}
\section{ftype}{
 Ca
}

\examples{
hgch_bar_hor_top_Ca(sampleData("Ca-Nu",nrow = 10))
}
