% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_hor_top_Cat}
\alias{hgch_bar_hor_top_Cat}
\title{Ordered horizontal bar}
\usage{
hgch_bar_hor_top_Cat(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, reverse = TRUE,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Ordered horizontal bar
}
\section{ctypes}{

Cat
}

\examples{
hgch_bar_hor_top_Cat(sampleData("Cat", nrow = 10))
}
