% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_hor_top_CatNum}
\alias{hgch_bar_hor_top_CatNum}
\title{Ordered horizontal bar}
\usage{
hgch_bar_hor_top_CatNum(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, reverse = TRUE,
  aggregate = "mean", theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Ordered horizontal bar
}
\section{ctypes}{

Cat-Num
}

\examples{
hgch_bar_hor_top_CatNum(sampleDatta("Cat-Num",nrow = 10))
}
