% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_hor_YeaNum}
\alias{hgch_bar_hor_YeaNum}
\title{Horizontal bar}
\usage{
hgch_bar_hor_YeaNum(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregation = "sum", color_title = "black", theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal bar
}
\section{ctypes}{

Yea-Num
}

\examples{
hgch_bar_hor_YeaNum(sampleData("Yea-Num", nrow = 10))
}
