% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_NumP}
\alias{hgch_bar_NumP}
\alias{hgch_bar_ver_NumP}
\title{Vertical bar}
\usage{
hgch_bar_ver_NumP(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical bar
}
\section{ftypes}{
 NumP
}

\examples{
hgch_bar_ver_NumP(sampleData("NumP",nrow = 10))
}
