% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_Oca}
\alias{hgch_bar_Oca}
\title{Bar (ordered categories)}
\usage{
hgch_bar_Oca(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  dropNa = FALSE, order = NULL, orientation = "ver", sort = "no",
  sliceN = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Oca
}

\examples{
hgch_bar_Oca(sampleData("Cat", nrow = 10))
}
