% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_OcaNum}
\alias{hgch_bar_OcaNum}
\title{Bar (ordered categories, numbers)}
\usage{
hgch_bar_OcaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  agg = "sum", dropNa = FALSE, order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among categories
}
\section{ctypes}{

Oca-Num
}

\examples{
hgch_bar_OcaNum(sampleData("Cat-Num", nrow = 10))
}
