% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-viz-opts.R
\name{hgch_bar_options}
\alias{hgch_bar_options}
\title{Bar options}
\arguments{
\item{bar_groupWidth}{Number, padding between each bar. Default is 0.1}
}
\description{
This page describes the bar options arguments
}
\keyword{internal}
