% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stack_CaCaNu}
\alias{hgch_bar_stack_CaCaNu}
\title{hgch_bar_stack_CaCaNu}
\usage{
hgch_bar_stack_CaCaNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, symbol = NULL, startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_stack_CaCaNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_bar_stack_CaCaNu(sampleData("Ye-Nu-Nu",nrow = 10))
}

