% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stack_CaDaNu}
\alias{hgch_bar_stack_CaDaNu}
\title{hgch_bar_stack_CaDaNu}
\usage{
hgch_bar_stack_CaDaNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, symbol = NULL, startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_stack_CaDaNu
}
\section{ftype}{
 Ca-Ca-Nu
}
\examples{
hgch_bar_stack_CaDaNu(sampleData("Ca-Ca-Nu",nrow = 10))
}

