% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_100_CatCatNum}
\alias{hgch_bar_stacked_100_CatCatNum}
\title{100% stacked bar (categories, ordered categories, numbers)}
\usage{
hgch_bar_stacked_100_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  orientation = "ver", sliceN = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among 100% stacked categories
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bar_stacked_100_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
