% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_100_hor_CatDatNum}
\alias{hgch_bar_stacked_100_hor_CatDatNum}
\title{Horizontal 100 stacked bar}
\usage{
hgch_bar_stacked_100_hor_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  aggregation = "sum", startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal 100 stacked bar
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
hgch_bar_stacked_100_hor_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
