% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_100_ver_CaCa}
\alias{hgch_bar_stacked_100_ver_CaCa}
\title{hgch_bar_stacked_100_ver_CaCa}
\usage{
hgch_bar_stacked_100_ver_CaCa(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_stacked_100_ver_CaCa
}
\section{ftype}{
 Ca-Ca
}

\examples{
hgch_bar_stacked_100_ver_CaCa(sampleData("Ca-Ca",nrow = 10))
}
