% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_100_ver_CaDaNu}
\alias{hgch_bar_stacked_100_ver_CaDaNu}
\title{Vertical 100% stacked bar}
\usage{
hgch_bar_stacked_100_ver_CaDaNu(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical 100% stacked bar
}
\section{ftype}{
 Ca-Da-Nu
}

\examples{
hgch_bar_stacked_100_ver_CaDaNu(sampleData("Ca-Da-Nu",nrow = 10))
}
