% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_CatOca}
\alias{hgch_bar_stacked_CatOca}
\title{Stacked bar (categories, ordered categories)}
\usage{
hgch_bar_stacked_CatOca(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, dropNa = FALSE, order = NULL, orientation = "ver",
  sort = "no", sliceN = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare stacked categories
}
\section{ctypes}{

Cat-Oca
}

\examples{
hgch_bar_stacked_CatOca(sampleData("Cat-Cat", nrow = 10))
}
