% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_hor_CatCatNum}
\alias{hgch_bar_stacked_hor_CatCatNum}
\title{Horizontal stacked bar}
\usage{
hgch_bar_stacked_hor_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal stacked bar
}
\section{ftype}{
 Cat-Cat-Num
}
\examples{
hgch_bar_stacked_hor_CatCatNum(sampleDatta("Cat-Cat-Num",nrow = 10))
}

