% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_ver_CatCat}
\alias{hgch_bar_stacked_ver_CatCat}
\title{hgch_bar_stacked_ver_CatCat}
\usage{
hgch_bar_stacked_ver_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_stacked_ver_CatCat
}
\section{ftype}{

Cat-Cat
}

\examples{
hgch_bar_stacked_ver_CatCat(sampleDatta("Cat-Cat",nrow = 10))
}
