% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_ver_CatDatNum}
\alias{hgch_bar_stacked_ver_CatDatNum}
\title{Vertical stacked bar}
\usage{
hgch_bar_stacked_ver_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical stacked bar
}
\section{ftype}{
 Cat-Dat-Num
}

\examples{
hgch_bar_stacked_ver_CatDatNum(sampleDatta("Cat-Dat-Num",nrow = 10))
}
