% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_Ca}
\alias{hgch_bar_ver_Ca}
\title{Vertical bar}
\usage{
hgch_bar_ver_Ca(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = "count", theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical bar
}
\section{ftype}{
 Ca
}

\examples{
hgch_bar_ver_Ca(sampleData("Ca",nrow = 10))
}
