% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_CaNu}
\alias{hgch_bar_ver_CaNu}
\title{Vertical bar}
\usage{
hgch_bar_ver_CaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = NULL,
  back_color = "white", color_title = "black", startAtZero = FALSE,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical bar
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_bar_ver_CaNu(sampleData("Ca-Nu",nrow = 10))
}
