% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_Cat}
\alias{hgch_bar_ver_Cat}
\title{Vertical bar}
\usage{
hgch_bar_ver_Cat(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = "count", theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical bar
}
\section{ftype}{

Cat
}
\examples{

hgch_bar_ver_Cat(sampleDatta("Cat",nrow = 10))

}

