% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_NumNumNumNum}
\alias{hgch_bar_ver_NumNumNumNum}
\title{Vertical bar}
\usage{
hgch_bar_ver_NumNumNumNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  aggregation = "sum", startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical bar
}
\section{ctypes}{

Num-Num-Num-Num
}

\examples{
hgch_bar_ver_NumNumNumNum(sampleData("Num-Num-Num-Num", nrow = 10))
}
