% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatNum.R
\name{hgch_bar_Yea}
\alias{hgch_bar_Yea}
\title{Bar Chart Yea}
\usage{
hgch_bar_Yea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Yea
}

\examples{
hgch_bar_Yea(sample_data("Yea", nrow = 10))
}
