% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatCatNum.R
\name{hgch_bar_YeaCatNum}
\alias{hgch_bar_YeaCatNum}
\title{bar Chart Yea Cat Num}
\usage{
hgch_bar_YeaCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bar Chart Yea Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bar_YeaCatNum(sample_data("Yea-Cat-Num", nrow = 10))
}
