% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatNum.R
\name{hgch_bar_YeaNum}
\alias{hgch_bar_YeaNum}
\title{Bar Chart Yea Num}
\usage{
hgch_bar_YeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Yea
}

\examples{
hgch_bar_YeaNum(sample_data("Yea-Num", nrow = 10))
}
