% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_YeNu}
\alias{hgch_bar_YeNu}
\title{hgch_bar_YeNu}
\usage{
hgch_bar_YeNu(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_YeNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_bar_YeNu(sampleData("Ye-Nu",nrow = 10))
}

