% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_bubble_CatNum}
\alias{hgch_bubble_CatNum}
\title{Bubble}
\usage{
hgch_bubble_CatNum(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Bubble
}
\section{ftype}{
 Cat-Num
}

\examples{
hgch_bubble_CatNum(sampleData("Cat-Num",nrow = 10))
}
