% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatNum.R
\name{hgch_bubbles_Cat}
\alias{hgch_bubbles_Cat}
\title{Bubbles Chart Cat}
\usage{
hgch_bubbles_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Bubbles Chart Cat
}
\section{ctypes}{

Cat
}

\examples{
hgch_bubbles_Cat(sampleData("Cat", nrow = 10))
}
