% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatCat.R
\name{hgch_bubbles_CatCat}
\alias{hgch_bubbles_CatCat}
\title{Bubbles Chart Cat Cat Numeric}
\usage{
hgch_bubbles_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat, Cat-Yea-Num
}

\examples{
hgch_bubbles_CatCat(sampleData("Cat-Cat", nrow = 10))
}
