% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{hgch_bubbles_CatCatNum}
\alias{hgch_bubbles_CatCatNum}
\title{Bubbles Cat Cat Num}
\usage{
hgch_bubbles_CatCatNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{}
}
\description{
Bubbles Cat Cat Num
}
