% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatNum.R
\name{hgch_bubbles_CatNum}
\alias{hgch_bubbles_CatNum}
\title{Bubbles Chart Cat Numeric}
\usage{
hgch_bubbles_CatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Num, Yea-Num
}

\examples{
hgch_bubbles_CatNum(sampleData("Cat-Num", nrow = 10))
}
