% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatCat.R
\name{hgch_bubbles_CatYea}
\alias{hgch_bubbles_CatYea}
\title{bubbles Chart Cat Yea}
\usage{
hgch_bubbles_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bubbles Chart Cat Yea
}
\section{ctypes}{

Cat-Yea
}

\examples{
hgch_bubbles_CatYea(sample_data("Cat-Yea", nrow = 10))
}
