% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatCatNum.R
\name{hgch_bubbles_CatYeaNum}
\alias{hgch_bubbles_CatYeaNum}
\title{bubbles Chart Cat Yea Num}
\usage{
hgch_bubbles_CatYeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bubbles Chart Cat Yea Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bubbles_CatYeaNum(sample_data("Cat-Yea-Num", nrow = 10))
}
