% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-viz-opts.R
\name{hgch_bubbles_options}
\alias{hgch_bubbles_options}
\title{Bubbles options}
\arguments{
\item{bubble_min}{The smallest area allowed in the bubbles plot. Default value is 30.}

\item{bubble_max}{The largest area allowed in the bubbles plot. Default value is 120.}

\item{bubble_opacity}{Number identicating color intensity of bubbles,
where 0 is transparent and 1 maximum intensity. Default is 1."}
}
\description{
This page describes the bar options arguments
}
\keyword{internal}
