% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_Cat.R
\name{hgch_bubbles_Yea}
\alias{hgch_bubbles_Yea}
\title{Bubbles chart Yea}
\usage{
hgch_bubbles_Yea(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read \code{\link[dsvizopts]{chart_viz_options}} a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_bubbles_options]{bubbles-options}}> which specifically contains the additional arguments
that work only for this type of chart.}
}
\description{
\code{hgch_bubbles_Yea()} Create a highcharter bubbles plot based on a particular data type.
In this case, you can load data with only one \strong{Year column} or be sure of
the first column of the dataframe its a year column since it will be done a
counting the years of this column
}
\section{Ftype}{

Yea
}

\examples{
data <- sample_data("Yea", n = 30)
hgch_bubbles_Yea(data)

# data with more of one column
data <- sample_data("Yea-Num-Dat-Yea-Yea", n = 30)
hgch_bubbles_Yea(data)

# calculate percentage
hgch_bubbles_Yea(data, percentage = TRUE)

# You can call the count and percentage in the tooltip plot
data \%>\%
hgch_bubbles_Yea(tooltip = "Count: {Count} <br/> Percentage: {\%}\%")

}
\seealso{
Other Yea plots: 
\code{\link{hgch_area_Yea}()},
\code{\link{hgch_bar_Yea}()},
\code{\link{hgch_donut_Yea}()},
\code{\link{hgch_line_Yea}()},
\code{\link{hgch_pie_Yea}()},
\code{\link{hgch_treemap_Yea}()}
}
\concept{Yea plots}
