% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatCatNum.R
\name{hgch_bubbles_YeaCatNum}
\alias{hgch_bubbles_YeaCatNum}
\title{bubbles Chart Yea Cat Num}
\usage{
hgch_bubbles_YeaCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bubbles Chart Yea Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bubbles_YeaCatNum(sample_data("Yea-Cat-Num", nrow = 10))
}
