% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatNum.R
\name{hgch_bubbles_YeaNum}
\alias{hgch_bubbles_YeaNum}
\title{bubbles Chart Yea Num}
\usage{
hgch_bubbles_YeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bubbles Chart Yea Num
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_bubbles_YeaNum(sample_data("Yea-Num", nrow = 10))
}
