% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_choropleth_GnmNum.R
\name{hgch_choropleth_GnmNum}
\alias{hgch_choropleth_GnmNum}
\title{map (gnm, numbers)}
\usage{
hgch_choropleth_GnmNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
map (gnm, numbers)
}
