% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_circular_bar_CaNu}
\alias{hgch_circular_bar_CaNu}
\title{Circular bar}
\usage{
hgch_circular_bar_CaNu(data, title = NULL, subtitle = NULL,
  caption = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Circular bar
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_circular_bar_CaNu(sampleData("Ca-Nu",nrow = 10))
}
