% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_circular_bar_CatNum}
\alias{hgch_circular_bar_CatNum}
\title{Circular bar}
\usage{
hgch_circular_bar_CatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Circular bar
}
\section{ftype}{
 Cat-Num
}
\examples{
hgch_circular_bar_CatNum(sampleDatta("Cat-Num",nrow = 10))
}

