% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_donut_Cat.R
\name{hgch_donut_Cat}
\alias{hgch_donut_Cat}
\title{donut Chart Cat}
\usage{
hgch_donut_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat, Yea-Num
}

\examples{
hgch_donut_Cat(sample_data("Cat", nrow = 10))
}
