% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_donut_CatNum.R
\name{hgch_donut_Yea}
\alias{hgch_donut_Yea}
\title{donut Chart Yea}
\usage{
hgch_donut_Yea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
donut Chart Yea
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_donut_Yea(sample_data("Yea", nrow = 10))
}
