% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_donut_CatNum.R
\name{hgch_donut_YeaNum}
\alias{hgch_donut_YeaNum}
\title{donut Chart Yea Num}
\usage{
hgch_donut_YeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
donut Chart Yea Num
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_donut_YeaNum(sample_data("Yea-Num", nrow = 10))
}
