% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_dumbbell_CatNumNum.R
\name{hgch_dumbbell_CatCat}
\alias{hgch_dumbbell_CatCat}
\title{Dumbbell Cat Cat}
\usage{
hgch_dumbbell_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Dumbbell Cat Cat
}
\section{}{
NA
}

\examples{
hgch_dumbbell_CatCat(sample_data("Cat-Cat", nrow = 1000))
}
