% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_dumbbell_CatNumNum.R
\name{hgch_dumbbell_CatCatNum}
\alias{hgch_dumbbell_CatCatNum}
\title{Dumbbell Cat Cat Num}
\usage{
hgch_dumbbell_CatCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Dumbbell Cat Cat Num
}
\section{}{
NA
}

\examples{
hgch_dumbbell_CatCatNum(sample_data("Cat-Cat-Num", nrow = 1000))
}
