% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_dumbbell_CatNumNum.R
\name{hgch_dumbbell_CatNumNum}
\alias{hgch_dumbbell_CatNumNum}
\title{Dumbbell Cat Num Num}
\usage{
hgch_dumbbell_CatNumNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Dumbbell Cat Num Num
}
\section{}{
NA
}

\examples{
hgch_dumbbell_CatNumNum(sample_data("Cat-Num-Num", nrow = 1000) \%>\% group_by_at(1) \%>\% summarise_all(mean))
}
