% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{hgch_funnel_CaNu}
\alias{hgch_funnel_CaNu}
\title{hgch_funnel_CaNu}
\usage{
hgch_funnel_CaNu(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_funnel_CaNu
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_funnel_CaNu(sampleData("Ca-Nu",nrow = 10))
}
