% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bla.R
\name{hgch_funnel_CatNum}
\alias{hgch_funnel_CatNum}
\title{Funnel}
\usage{
hgch_funnel_CatNum(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Funnel
}
\section{ctypes}{

Cat-Num
}

\examples{

hgch_funnel_CatNum(sampleData("Cat-Num",nrow = 10))

}
