% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{hgch_heatmap_matrix_CatCatNum}
\alias{hgch_heatmap_matrix_CatCatNum}
\title{Heatmap matrix}
\usage{
hgch_heatmap_matrix_CatCatNum(
  data,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  xAxisTitle = NULL,
  yAxisTitle = NULL,
  sort = "no",
  theme = NULL,
  export = FALSE,
  ...
)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Heatmap matrix
}
\section{ctypes}{

CatCatNum
}

\examples{
hgch_heatmap_matrix_CatCatNum(sampleData("Cat-Cat-Num", nrow = 20))
}
