% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_line_CaNu}
\alias{hgch_line_CaNu}
\title{hgch_line_CaNu}
\usage{
hgch_line_CaNu(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  symbol = NULL, startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_line_CaNu
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_line_CaNu(sampleData("Ca-Nu",nrow = 10))
}

