% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_Cat}
\alias{hgch_line_Cat}
\title{Lines (categories)}
\usage{
hgch_line_Cat(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_line_Cat(sampleData("Cat", nrow = 10))
}
