% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatCat}
\alias{hgch_line_CatCat}
\title{Line (categories, ordered categories, numbers)}
\usage{
hgch_line_CatCat(
  data,
  agg_text = NULL,
  allow_point = FALSE,
  background = "#ffffff",
  border_color = "#CCCCCC",
  border_width = 1,
  caption = NULL,
  click_function = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_opacity = 0.7,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  drop_na_legend = FALSE,
  export = FALSE,
  fill_opacity = 0.5,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = " ",
  label_wrap = 12,
  label_wrap_legend = 12,
  lang = "es",
  legend_position = "center",
  legend_show = TRUE,
  marks = c(".", ","),
  n_digits = NULL,
  null_color = "#f7f7f7",
  order1 = NULL,
  order2 = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  spline = FALSE,
  start_zero = TRUE,
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = " ",
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat
}

\examples{
hgch_line_CatCat(sampleData("Cat-Cat-Num", nrow = 10))
}
