% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_CatDatNum.R
\name{hgch_line_CatDatNum}
\alias{hgch_line_CatDatNum}
\title{line Chart Cat Dat Numeric}
\usage{
hgch_line_CatDatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
line Chart Cat Dat Numeric
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
hgch_line_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
