% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatNumP}
\alias{hgch_line_CatNumP}
\title{Line (ordered category, n numbers)}
\usage{
hgch_line_CatNumP(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Cat-NumP
}

\examples{
hgch_line_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
