% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatOcaNum}
\alias{hgch_line_CatOcaNum}
\title{Line (categories, ordered categories, numbers)}
\usage{
hgch_line_CatOcaNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Oca-Num
}

\examples{
hgch_line_CatOcaNum(sampleData("Cat-Oca-Num", nrow = 10))
}
