% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_CatYeaNum.R
\name{hgch_line_CatYea}
\alias{hgch_line_CatYea}
\title{line Chart Cat Yea}
\usage{
hgch_line_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
line Chart Cat Yea
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_line_CatYea(sample_data("Cat-Yea", nrow = 10))
}
