% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_CatYeaNum.R
\name{hgch_line_CatYeaNum}
\alias{hgch_line_CatYeaNum}
\title{line Chart Cat Cat Numeric}
\usage{
hgch_line_CatYeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num
}

\examples{
hgch_line_CatYeaNum(sample_data("Cat-Yea-Num", nrow = 100))
}
