% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_line_CaYeNu}
\alias{hgch_line_CaYeNu}
\title{hgch_line_CaYeNu}
\usage{
hgch_line_CaYeNu(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  symbol = NULL, startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_line_CaYeNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_line_CaYeNu(sampleData("Ye-Nu-Nu",nrow = 10))
}

