% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_line_cdn}
\alias{hgch_line_cdn}
\title{hgch_line_cdn}
\usage{
hgch_line_cdn(f, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  symbol = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_line_cdn
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_line_cdn(sampleData("Ca-Da-Nu",nrow = 10))
}

