% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_line_cyn}
\alias{hgch_line_cyn}
\title{hgch_line_cyn}
\usage{
hgch_line_cyn(f, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  symbol = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_line_cyn
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_line_cyn(sampleData("Ye-Nu-Nu",nrow = 10))
}

