% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_DatNum}
\alias{hgch_line_DatNum}
\title{Line (dates, numbers)}
\usage{
hgch_line_DatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare a quantities over time (Year-month-day)
}
\section{ctypes}{

Dat-Num
}

\examples{
hgch_line_DatNum(sampleData("Dat-Num", nrow = 10))
}
