% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_DatNumP}
\alias{hgch_line_DatNumP}
\title{Line (dates, n numbers)}
\usage{
hgch_line_DatNumP(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  agg = "sum", dropNa = FALSE, order = NULL, percentage = FALSE,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare n quantities over dates
}
\section{ctypes}{

Dat-NumP
}

\examples{
hgch_line_DatNumP(sampleData("Dat-NumP", nrow = 10))
}
