% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_Oca}
\alias{hgch_line_Oca}
\title{Line (ordered categories)}
\usage{
hgch_line_Oca(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL, yLineLabel = NULL,
  dropNa = FALSE, order = NULL, percentage = FALSE, startAtZero = FALSE,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Oca
}

\examples{
hgch_line_Oca(sampleData("Oca", nrow = 10))
}
