% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_OcaNum}
\alias{hgch_line_OcaNum}
\title{Line (ordered categories, numbers)}
\usage{
hgch_line_OcaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among categories
}
\section{ctypes}{

Oca-Num
}

\examples{
hgch_line_OcaNum(sampleData("Oca-Num", nrow = 10))
}
