% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_Yea.R
\name{hgch_line_Yea}
\alias{hgch_line_Yea}
\title{Line chart Yea Num}
\usage{
hgch_line_Yea(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read \code{\link[dsvizopts]{chart_viz_options}} a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_line_options]{line-options}}> which specifically contains the additional arguments
that work only for this type of chart.}
}
\description{
\code{hgch_line_Yea()} Create a highcharter line plot based on a particular data type.
In this case, you can load data with only one column, where this is a
\strong{Year column}, or be sure of the first column of the dataframe its a year column,
since it will be done a counting the categories of this column
}
\section{Ftype}{

Yea
}

\examples{
data <- sample_data("Yea", n = 300, rep = T)
hgch_line_Yea(data)

# data with more of one column
data <- sample_data("Yea-Dat-Yea-Yea", n = 30, rep = T)
hgch_line_Yea(data)

# calculate percentage
hgch_line_Yea(data, percentage = TRUE)

# numeric format
hgch_line_Yea(data, percentage = TRUE, format_sample_num = "1.234,")

# You can call the mean and percentage in the tooltip plot
num_name <- names(data)[2]
data \%>\%
hgch_line_Yea(tooltip = paste0("Average: {", num_name ,"} <br/> Percentage: {\%}\%"))

}
\seealso{
Other Yea plots: 
\code{\link{hgch_area_Yea}()},
\code{\link{hgch_bar_Yea}()},
\code{\link{hgch_bubbles_Yea}()},
\code{\link{hgch_donut_Yea}()},
\code{\link{hgch_pie_Yea}()},
\code{\link{hgch_treemap_Yea}()}
}
\concept{Yea plots}
