% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_CatYeaNum.R
\name{hgch_line_YeaCatNum}
\alias{hgch_line_YeaCatNum}
\title{line Chart Yea Cat Num}
\usage{
hgch_line_YeaCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
line Chart Yea Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_line_YeaCatNum(sample_data("Yea-Cat-Num", nrow = 10))
}
