% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_YeaNumP}
\alias{hgch_line_YeaNumP}
\title{Line (years, n numbers)}
\usage{
hgch_line_YeaNumP(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  percentage = FALSE, startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare n quantities over years
}
\section{ctypes}{

Yea-NumP
}

\examples{
hgch_line_YeaNumP(sampleData("Yea-NumP", nrow = 10))
}
