% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_YeNu}
\alias{hgch_line_YeNu}
\title{hgch_line_YeNu}
\usage{
hgch_line_YeNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_line_YeNu
}
\section{ftype}{
 Ca-Ye-Nu
}

\examples{
hgch_line_YeNu(sampleData("Ye-Nu",nrow = 10))
}
