% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_bubbles_CatGltGlnNum}
\alias{hgch_map_bubbles_CatGltGlnNum}
\title{map (gnm, numbers)}
\usage{
hgch_map_bubbles_CatGltGlnNum(
  data = NULL,
  mapName = "countries/co/co-all",
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\description{
map (gnm, numbers)
}
