% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_bubbles_latinAmerican_GcdNum}
\alias{hgch_map_bubbles_latinAmerican_GcdNum}
\title{Bubble latam map}
\usage{
hgch_map_bubbles_latinAmerican_GcdNum(data, title = NULL, subtitle = NULL,
  geoinfoPath = NULL, geoCodeVar = NULL, geoNameVar = NULL,
  theme = NULL, col_bur = NULL, export = FALSE, leg_pos = "bottom",
  leg_col = "#505053", leg_alg = "left", back_color = "#CDD2D4", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Bubble latam map
}
\section{ftype}{
 Gcd-Num
}

\examples{
hgch_map_bubbles_latinAmerican_GcdNum(sampleData("Gcd-Num",nrow = 10))
}
