% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_bubbles_latinAmerican_GcdNumNum}
\alias{hgch_map_bubbles_latinAmerican_GcdNumNum}
\title{Bubble latam map}
\usage{
hgch_map_bubbles_latinAmerican_GcdNumNum(data, title = NULL,
  subtitle = NULL, geoinfoPath = NULL, geoCodeVar = NULL,
  geoNameVar = NULL, theme = NULL, col_bone = NULL, col_btwo = NULL,
  export = FALSE, leg_pos = "bottom", leg_col = "#505053",
  leg_alg = "left", back_color = "#CDD2D4", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Bubble latam map
}
\section{ftype}{
 Gcd-Num-Num
}

\examples{
hgch_map_bubbles_latinAmerican_GcdNumNum(sampleData("Gcd-Num-Num",nrow = 10))
}
