% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_bubbles_latinAmerican_GeNuNu}
\alias{hgch_map_bubbles_latinAmerican_GeNuNu}
\title{hgch_map_bubbles_latinAmerican_GeNuNu}
\usage{
hgch_map_bubbles_latinAmerican_GeNuNu(data, title = NULL, subtitle = NULL,
  geoinfoPath = NULL, geoCodeVar = NULL, geoNameVar = NULL,
  theme = NULL, export = FALSE, leg_pos = "bottom", leg_col = "#505053",
  leg_alg = "left", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_map_bubbles_latinAmerican_GeNuNu
}
\section{ftype}{
 Ca, Ca0Nu
}

\examples{
hgch_map_bubbles_latinAmerican_GeNuNu(sampleData("Ca",nrow = 10))
}
