% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_bubbles_world_GcdNum}
\alias{hgch_map_bubbles_world_GcdNum}
\title{Bubble world map}
\usage{
hgch_map_bubbles_world_GcdNum(data, title = NULL, subtitle = NULL,
  geoinfoPath = NULL, geoCodeVar = NULL, geoNameVar = NULL,
  theme = NULL, export = FALSE, leg_pos = "bottom", leg_col = "#505053",
  leg_alg = "left", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Bubble world map
}
\section{ftype}{
 Gcd-Num
}

\examples{
hgch_map_bubbles_world_GcdNum(sampleData("Gcd-Num",nrow = 10))
}
