% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_bubbles_world_GeNu}
\alias{hgch_map_bubbles_world_GeNu}
\title{hgch_map_bubbles_world_GeNu}
\usage{
hgch_map_bubbles_world_GeNu(data, title = NULL, subtitle = NULL,
  geoinfoPath = NULL, geoCodeVar = NULL, geoNameVar = NULL,
  theme = NULL, export = FALSE, leg_pos = "bottom", leg_col = "#505053",
  leg_alg = "left", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_map_bubbles_world_GeNu
}
\section{ftype}{
 Ge-Nu
}

\examples{
hgch_map_bubbles_world_GeNu(sampleData("Ge-Nu",nrow = 10))
}
