% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_choro_world_GcdNum}
\alias{hgch_map_choro_world_GcdNum}
\title{Choropleth's world map}
\usage{
hgch_map_choro_world_GcdNum(data, title = NULL, subtitle = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, minColor = "#E63917",
  maxColor = "#18941E", aggregate = "count", theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Choropleth's world map
}
\section{ftype}{
 Gcd-Num
}

\examples{
hgch_map_choro_world_GcdNum(sampleData("Gcd-Num",nrow = 10))
}
