% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_choro_world_GeNu}
\alias{hgch_map_choro_world_GeNu}
\title{hgch_map_choro_world_GeNu}
\usage{
hgch_map_choro_world_GeNu(data, title = NULL, subtitle = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, minColor = "#E63917",
  maxColor = "#18941E", aggregate = "count", theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_map_choro_world_GeNu
}
\section{ftype}{
 Ca
}
\examples{
hgch_map_choro_world_GeNu(sampleData("Ca",nrow = 10))
}

