% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_choropleth_Gnm}
\alias{hgch_map_choropleth_Gnm}
\title{map (gnm)}
\usage{
hgch_map_choropleth_Gnm(data = NULL, mapName = "countries/co/co-all",
  opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
map (gnm)
}
