% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{hgch_map_choropleth_GnmNum}
\alias{hgch_map_choropleth_GnmNum}
\title{map (gnm, numbers)}
\usage{
hgch_map_choropleth_GnmNum(data = NULL,
  mapName = "countries/co/co-all", opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
map (gnm, numbers)
}
