% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_mapbubbles_GnmNum.R
\name{hgch_mapbubbles_GltGlnNum}
\alias{hgch_mapbubbles_GltGlnNum}
\title{map (gnm, numbers)}
\usage{
hgch_mapbubbles_GltGlnNum(data = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
map (gnm, numbers)
}
