% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_multilines.}
\alias{hgch_multilines.}
\alias{hgch_multilines_YeNuP}
\alias{hgch_multilines.}
\alias{hgch_slope_CaYeNu}
\title{hgch_multilines_YeNuP
Multilines}
\usage{
hgch_multilines_YeNuP(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)

hgch_slope_CaYeNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}

\item{x}{A data.frame}
}
\value{
highcharts viz

highcharts viz
}
\description{
hgch_multilines_YeNuP
Multilines

hgch_slope_CaYeNu
Slope
}
\section{ftypes}{
 Ye-NuP

 Ca-Ye-Nu
}

\examples{
hgch_multilines_YeNuP(sampleData("Ye-Nu-Nu",nrow = 10))
}
