% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_multilines_YeaNumNumNumNum}
\alias{hgch_multilines_YeaNumNumNumNum}
\title{Grouped line}
\usage{
hgch_multilines_YeaNumNumNumNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  aggregation = "sum", startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Grouped line
}
\section{ctypess}{
 Yea-Num-Num-Num-Num
}

\examples{

hgch_multilines_YeaNumNumNumNum(sampleData("Yea-Num-Num-Num-Num", nrow = 10))

}
