% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_multilines_YeNuP}
\alias{hgch_multilines_YeNuP}
\title{Grouped line}
\usage{
hgch_multilines_YeNuP(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Grouped line
}
\section{ftypes}{
 Ye-NuP
}

\examples{
hgch_multilines_YeNuP(sampleData("Ye-NuP",nrow = 10))
}
