% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_pie_Ca}
\alias{hgch_pie_Ca}
\title{hgch_pie_Ca}
\usage{
hgch_pie_Ca(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  sort = "no", aggregate = "count", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_pie_Ca
}
\section{ftype}{
 Ca
}
\examples{
hgch_pie_Ca(sampleData("Ca",nrow = 10))
}

