% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_pie_CaNu}
\alias{hgch_pie_CaNu}
\title{Pie}
\usage{
hgch_pie_CaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, back_color = "white",
  leg_op = TRUE, color_title = "black", sort = "no", aggregate = "sum",
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Pie
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_pie_CaNu(sampleData("Ca-Nu",nrow = 10))
}
