% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{hgch_pie_CatNum}
\alias{hgch_pie_CatNum}
\title{Pie (categories, numbers)}
\usage{
hgch_pie_CatNum(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing quantities among categories
}
\section{ctypes}{

Cat-Num
}

\examples{
hgch_pie_CatNum(sampleData("Cat-Num", nrow = 10))
}
