% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_pie_Cat.R
\name{hgch_pie_Yea}
\alias{hgch_pie_Yea}
\title{Pie chart Yea}
\usage{
hgch_pie_Yea(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{<\code{\link[=hgch_viz_options]{chart-options}}> Options to configure your hgchmagic plots}
}
\description{
\code{hgch_pie_Yea()} Create a highcharter pie plot based on a particular data type.
In this case, you can load data with only one \strong{Year column} or be sure of
the first column of the dataframe its a year column since it will be done a
counting the years of this column
}
\section{Ftype}{

Yea
}

\examples{
data <- sample_data("Yea", n = 30)
hgch_pie_Yea(data)

# data with more of one column
data <- sample_data("Yea-Num-Dat-Yea-Yea", n = 30)
hgch_pie_Yea(data)

# calculate percentage
hgch_pie_Yea(data, percentage = TRUE)

# You can call the count and percentage in the tooltip plot
data \%>\%
hgch_pie_Yea(tooltip = "Count: {Count} <br/> Percentage: {\%}\%")
}
\seealso{
Other Yea plots: 
\code{\link{hgch_bar_Yea}()},
\code{\link{hgch_line_Yea}()},
\code{\link{hgch_treemap_Yea}()}
}
\concept{Yea plots}
