% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_pie_CatNum.R
\name{hgch_pie_YeaNum}
\alias{hgch_pie_YeaNum}
\title{pie Chart Yea Num}
\usage{
hgch_pie_YeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
pie Chart Yea Num
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_pie_YeaNum(sample_data("Yea-Num", nrow = 10))
}
