% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{hgch_polarbar_grouped_CatOcaNum}
\alias{hgch_polarbar_grouped_CatOcaNum}
\title{Polar bar (categories, ordered categories, numbers)}
\usage{
hgch_polarbar_grouped_CatOcaNum(data, title = NULL, subtitle = NULL,
  caption = NULL, yLine = NULL, agg = "sum", dropNa = FALSE,
  order = NULL, percentage = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare grouped quantities among two categories
}
\section{ctypes}{

Cat-Oca-Num
}

\examples{
hgch_polarbar_grouped_CatOcaNum(sampleData("Cat-Cat-Num", nrow = 10))
}
