% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{hgch_polarbar_Oca}
\alias{hgch_polarbar_Oca}
\title{Polar bar (ordered categories)}
\usage{
hgch_polarbar_Oca(data, title = NULL, subtitle = NULL,
  caption = NULL, yLine = NULL, dropNa = FALSE, order = NULL,
  sort = "no", sliceN = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing counts of categories
}
\section{ctypes}{

Oca
}

\examples{
hgch_polarbar_Oca(sampleData("Cat", nrow = 10))
}
