% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_polarcolumn_CaNu}
\alias{hgch_polarcolumn_CaNu}
\title{Polar bar}
\usage{
hgch_polarcolumn_CaNu(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Polar bar
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_polarcolumn_CaNu(sampleData("Ca-Nu",nrow = 10))
}
