% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_polarcolumn_Cat}
\alias{hgch_polarcolumn_Cat}
\title{Polar bar}
\usage{
hgch_polarcolumn_Cat(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Polar bar
}
\section{ftype}{
 Cat
}

\examples{
hgch_polarcolumn_Cat(sampleData("Cat",nrow = 10))
}
