% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_polarcolumn_CatNum}
\alias{hgch_polarcolumn_CatNum}
\title{Polar bar}
\usage{
hgch_polarcolumn_CatNum(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Polar bar
}
\section{ftype}{
 Cat-Num
}

\examples{
hgch_polarcolumn_CatNum(sampleData("Cat-Num",nrow = 10))
}
