% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{hgch_pyramid_CaNu}
\alias{hgch_pyramid_CaNu}
\title{hgch_pyramid_CaNu}
\usage{
hgch_pyramid_CaNu(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_pyramid_CaNu
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_pyramid_CaNu(sampleData("Ca-Nu",nrow = 10))
}

