% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{hgch_pyramid_CatNum}
\alias{hgch_pyramid_CatNum}
\title{Pyramid}
\usage{
hgch_pyramid_CatNum(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Pyramid
}
\section{ctypes}{

Cat-Num
}

\examples{

hgch_pyramid_CatNum(sampleData("Cat-Num",nrow = 10))

}
