% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_radar_Ca}
\alias{hgch_radar_Ca}
\title{hgch_radar_Ca}
\usage{
hgch_radar_Ca(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "mean",
  export = FALSE, theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_radar_Ca
}
\section{ftype}{
 Ca
}

\examples{
hgch_radar_Ca(sampleData("Ca", nrow = 10))
}
