% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_radar_CaNu}
\alias{hgch_radar_CaNu}
\title{hgch_radar_CaNu}
\usage{
hgch_radar_CaNu(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  sort = "no", aggregate = "mean", export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_radar_CaNu
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_radar_CaNu(sampleData("Ca-Nu",nrow = 10))
}

