% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_radar_Cat}
\alias{hgch_radar_Cat}
\title{hgch_radar_Cat}
\usage{
hgch_radar_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "mean",
  export = FALSE, theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_radar_Cat
}
\section{ftype}{

Cat
}
\examples{

hgch_radar_Cat(sampleData("Cat", nrow = 10))

}

