% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_radar_CatNum}
\alias{hgch_radar_CatNum}
\title{hgch_radar_CatNum}
\usage{
hgch_radar_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "mean",
  export = FALSE, theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_radar_CatNum
}
\section{ftype}{

Cat-Num
}

\examples{

hgch_radar_CatNum(sampleData("Cat-Num",nrow = 10))

}
