% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_radar_CatNumNum}
\alias{hgch_radar_CatNumNum}
\title{hgch_radar_CatNumNum}
\usage{
hgch_radar_CatNumNum(data, title = NULL, subtitle = NULL, caption = NULL,
  sort = "no", aggregate = "mean", export = FALSE, theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_radar_CatNumNum
}
\section{ctypes}{

Cat-Num-Num
}

\examples{

hgch_radar_CatNumNum(sampleData("Cat-Num-Num", nrow = 10))

}
