% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_radar_CatNumNumNumNum}
\alias{hgch_radar_CatNumNumNumNum}
\title{hgch_radar_CatNumNumNumNum}
\usage{
hgch_radar_CatNumNumNumNum(data, title = NULL, subtitle = NULL,
  caption = NULL, sort = "no", aggregate = "mean", export = FALSE,
  theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_radar_CatNumNumNumNum
}
\section{ctypes}{

Cat-Num-Num-Num-Num
}

\examples{

hgch_radar_CatNumNumNumNum(sampleData("Cat-Num-Num-Num-Num", nrow = 10))

}
