% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{hgch_radar_OcaNum}
\alias{hgch_radar_OcaNum}
\title{Radar (ordered categories, numbers)}
\usage{
hgch_radar_OcaNum(
  data,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  yLine = NULL,
  dropNa = FALSE,
  agg = "sum",
  order = NULL,
  percentage = FALSE,
  sort = "no",
  sliceN = NULL,
  theme = NULL,
  export = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing quantities among categories
}
\section{ctypes}{

Oca-Num
}

\examples{
hgch_radar_OcaNum(sampleData("Cat-Num", nrow = 10))
}
