% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.R
\name{hgch_radar_OcaNumP}
\alias{hgch_radar_OcaNumP}
\title{Radar (ordered categories, n numbers)}
\usage{
hgch_radar_OcaNumP(
  data,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  yLine = NULL,
  dropNa = FALSE,
  agg = "sum",
  order = NULL,
  percentage = FALSE,
  theme = NULL,
  export = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing n quantities among categories
}
\section{ctypes}{

Oca-NumP
}

\examples{
hgch_radar_OcaNumP(sampleData("Cat-NumP", nrow = 10))
}
