% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_sankey.R
\name{hgch_sankey}
\alias{hgch_sankey}
\title{Highcharter sankey Chart}
\usage{
hgch_sankey(data, dic = NULL, var_cat = NULL, var_num = NULL, ...)
}
\arguments{
\item{data}{a data frame containing the variables to be plotted.}

\item{dic}{a named list, each element corresponding to a column name in \code{data} and defining a set of labels for that column.}

\item{var_cat}{the names of the categorical variables in the data frame}

\item{var_num}{a character vector with the name of numerical variable.}

\item{...}{additional arguments to be passed to \code{\link[dsvizopts]{dsviz_default_opts}}}
}
\value{
a Highcharter sankey chart.
}
\description{
Creates a sankey chart using Highcharter library.
}
\examples{
data <- ggplot2::diamonds
data <- data |>
  dplyr::group_by(cut, clarity) |>
  dplyr::summarise(total = sum(z, na.rm = T))

hgch_sankey(data, var_cat = c("cut", "clarity"), var_num = "total")

### Adding style to plot
test_theme <- list(
  theme = list(
    palette_colors = "#ffa92a",
    caption_align = "right",
    caption_family = "Roboto",
    caption_size = 15,
    caption_color = "#3b83b8",
    caption_weight = 700
  )
)

data <- dplyr::starwars
data <- data |>
  dplyr::select(hair_color, sex) |>
  tidyr::drop_na()

hgch_sankey_CatCat(data,
                   opts = test_theme,
                   caption = "theme caption")

}
\seealso{
\code{\link{data_draw}}
}
