% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_sankey_CatCat.R
\name{hgch_sankey_CatCat}
\alias{hgch_sankey_CatCat}
\title{Sankey chart Cat Cat}
\usage{
hgch_sankey_CatCat(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read <\code{\link[=hgch_viz_options]{chart-options}}> a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_sankey_options]{sankey-options}}> which specifically contains the additional arguments
that work only for this type of chart}
}
\description{
\code{hgch_sankey_CatCat()} Create a highcharter sankey plot based on a particular data type.
In this case, you can load data with only two columns, where the firts and second columns are
\strong{categoricals columns}, or be sure that firts two columns they meet this condition, since it
will be done a counting the categories of this columns.
}
\section{Ftype}{

Cat-Cat
}

\examples{
data <- sample_data("Cat-Cat", n = 30)
hgch_sankey_CatCat(data)

# Activate data labels
hgch_sankey_CatCat(data,
                       dataLabels_show = TRUE)

# data with more of one column
data <- sample_data("Cat-Cat-Num-Yea-Cat", n = 30)
hgch_sankey_CatCat(data)

}
\seealso{
Other Cat-Cat plots: 
\code{\link{hgch_treemap_CatCat}()}
}
\concept{Cat-Cat plots}
