% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_scatter}
\alias{hgch_scatter}
\alias{hgch_scatter_CaCaNuNu}
\alias{hgch_scatter_CaCaNuNuNu}
\title{hgch_scatter}
\usage{
hgch_scatter(f, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL)

hgch_scatter(f, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL)

hgch_scatter_CaCaNuNu(f, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL)

hgch_scatter_CaCaNuNuNu(f, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL)
}
\description{
hgch_scatter

hgch_scatter

hgch_scatter

hgch_scatter
}
\section{ftype}{
 Ca-Nu-Nu

 Ca-Nu-Nu

 Ca-Nu-Nu

 Ca-Nu-Nu
}

