% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CatCatNumNum}
\alias{hgch_scatter_CatCatNumNum}
\title{Scatter (categories, categories, numbers, numbers)}
\usage{
hgch_scatter_CatCatNumNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = NULL, verLine = NULL, verLineLabel = NULL,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Two numeric variables set in a cartesian plane with two categories
}
\section{ctypes}{

Cat-Cat-Num-Num
}

\examples{
hgch_scatter_CatCatNumNum(sampleData("Cat-Cat-Num-Num", nrow = 10))
}
