% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_scatter_CatDatNum.R
\name{hgch_scatter_CatDatNum}
\alias{hgch_scatter_CatDatNum}
\title{scatter Chart Cat Dat Numeric}
\usage{
hgch_scatter_CatDatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
scatter Chart Cat Dat Numeric
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
hgch_scatter_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
