% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_scatter_CatNumNum.R
\name{hgch_scatter_CatNumNum}
\alias{hgch_scatter_CatNumNum}
\title{scatter Chart Cat Num Numeric}
\usage{
hgch_scatter_CatNumNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
scatter Chart Cat Num Numeric
}
\section{ctypes}{

Cat-Num-Num
}

\examples{
hgch_scatter_CatNumNum(sample_data("Cat-Num-Num", nrow = 10))
}
