% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-viz-opts.R
\name{hgch_scatter_options}
\alias{hgch_scatter_options}
\title{Scatter Options}
\arguments{
\item{scatter_format_num_sample_x}{A specific format of the x-axis values of the graph.}

\item{scatter_format_num_sample_y}{A specific format of the y-axis values of the graph.}

\item{scatter_format_num_sample_size}{A specific format of the size values of bubble in the graph}

\item{scatter_prefix_x}{Character string to append after formatted x value.}

\item{scatter_suffix_x}{Character string to append before formatted x value.}

\item{scatter_prefix_y}{Character string to append after formatted y value.}

\item{scatter_suffix_y}{Character string to append before formatted y value.}

\item{scatter_regression}{Boolean. If TRUE, calculate a linear regression in machine learning with this you can analyze a relationship between variables}

\item{scatter_regression_color}{Color of the regression line}

\item{scatter_regression_equation}{Boolean. If TRUE, displays the formula used in the regression model}
}
\description{
This page describes the scatter options arguments
}
\keyword{internal}
