% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_slope_CatYeaNum}
\alias{hgch_slope_CatYeaNum}
\title{Slope}
\usage{
hgch_slope_CatYeaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Slope
}
\section{ftypes}{
 Cat-Yea-Num
}
\examples{

hgch_slope_CatYeaNum(sampleDatta('Cat-Yea-Num'))

}

