% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_slope_CaYeNu}
\alias{hgch_slope_CaYeNu}
\title{Slope}
\usage{
hgch_slope_CaYeNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Slope
}
\section{ftypes}{
 Ca-Ye-Nu
}

\examples{
hgch_slope_CaYeNu(sampleData('Ca-Ye-Nu'))
}
