% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_spider_CaNuNu}
\alias{hgch_spider_CaNuNu}
\title{hgch_spider_CaNuNu}
\usage{
hgch_spider_CaNuNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "mean",
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_spider_CaNuNu
}
\section{ftype}{
 Ca-Nu-Nu
}

\examples{
hgch_spider_CaNuNu(sampleData("Ca-Nu-Nu",nrow = 10))
}
