% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_spider_CatNum}
\alias{hgch_spider_CatNum}
\title{Radar}
\usage{
hgch_spider_CatNum(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Radar
}
\section{ftype}{
 Cat-Num
}
\examples{
hgch_spider_CatNum(sampleData("Cat-Num",nrow = 10))
}

