% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_spider_CatNumNum}
\alias{hgch_spider_CatNumNum}
\title{Radar}
\usage{
hgch_spider_CatNumNum(data, title = NULL, subtitle = NULL,
  Catption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = "mean", theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Radar
}
\section{ftype}{

Cat-Num-Num
}

\examples{

hgch_spider_CatNumNum(sampleData("Cat-Num-Num",nrow = 10))

}
