% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_spider_cn}
\alias{hgch_spider_cn}
\title{hgch_spider_cn}
\usage{
hgch_spider_cn(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  sort = "no", aggregate = "mean")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_spider_cn
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_spider_cn(sampleData("Ca-Nu",nrow = 10))
}

