% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_spider_cnn}
\alias{hgch_spider_cnn}
\title{hgch_spider_cnn}
\usage{
hgch_spider_cnn(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  sort = "no", aggregate = "mean", ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_spider_cnn
}
\section{ftype}{
 Ca-Nu-Nu
}
\examples{
hgch_spider_cnn(sampleData("Ca-Nu-Nu",nrow = 10))
}

