% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{hgch_stock_lines_DaNuP}
\alias{hgch_stock_lines_DaNuP}
\title{hgch_stock_lines_DaNuP}
\usage{
hgch_stock_lines_DaNuP(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, ...)
}
\arguments{
\item{data}{A dataframe}
}
\value{
highcharts viz
}
\description{
hgch_stock_lines_DaNuP
}
\section{ftypes}{
 Da-NuP
}
\examples{
hgch_stock_lines_DaNuP(sampleData("Da-NuP", nrow = 15))
}

