% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{hgch_stock_multilines_DaNu}
\alias{hgch_stock_multilines_DaNu}
\title{hgch_stock_multilines_DaNu}
\usage{
hgch_stock_multilines_DaNu(a, b, c = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, ...)
}
\arguments{
\item{a}{A dataframe}

\item{b}{A dataframe}

\item{c}{A dataframe}
}
\value{
highcharts viz
}
\description{
hgch_stock_multilines_DaNu
}
\section{ftypes}{
 Da-Nu
}
\examples{
hgch_stock_multilines_DaNu(sampleData("Da-Nu", nrow = 10), sampleData("Da-Nu", nrow = 18))
}

