% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{hgch_stock_ohlc_DaNu4}
\alias{hgch_stock_ohlc_DaNu4}
\title{hgch_stock_ohlc_DaNu4}
\usage{
hgch_stock_ohlc_DaNu4(a, type = "ohlc", title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, ...)
}
\arguments{
\item{a}{A dataframe}

\item{type}{Factor: "candlestick", "ohlc"}
}
\value{
highcharts viz
}
\description{
hgch_stock_ohlc_DaNu4
}
\section{ftypes}{
 Da-Nu-Nu-Nu-Nu
}
\examples{
hgch_stock_ohlc_DaNu4(sampleData("Da-Nu-Nu-Nu-Nu", nrow = 15))
}

