% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\docType{data}
\name{hgch_test_docs}
\alias{hgch_test_docs}
\title{hgch_test_docs
test}
\format{An object of class \code{NULL} of length 0.}
\usage{
hgch_test_docs
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
hgch_test_docs
test
}
\section{noftypes}{
 Ca,Ca-Nu
}

\examples{
add(1, 1)
}
\keyword{datasets}
