% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\docType{data}
\name{hgch_test_docs2}
\alias{hgch_test_docs2}
\title{hgch_test_docs2
test}
\format{An object of class \code{NULL} of length 0.}
\usage{
hgch_test_docs2
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
hgch_test_docs2
test
}
\section{ftypes}{
 Ca,Ca-Ca-Ca
}

\examples{
add(1, 1)
}
\keyword{datasets}
