% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap}
\alias{hgch_treemap}
\title{Highcharter treemap Chart}
\usage{
hgch_treemap(data, dic = NULL, var_cat = NULL, var_num = NULL, ...)
}
\arguments{
\item{data}{a data frame containing the variables to be plotted.}

\item{dic}{a named list, each element corresponding to a column name in \code{data} and defining a set of labels for that column.}

\item{var_cat}{a character vector with the names of categorical variables.}

\item{var_num}{a character vector with the names of numerical variables.}

\item{...}{additional arguments to be passed to \code{\link[dsvizopts]{dsviz_default_opts}}}
}
\value{
a Highcharter treemap chart.
}
\description{
Creates a treemap chart using Highcharter library.
}
\examples{
data(mtcars)
mtcars$cyl <- as.character(mtcars$cyl)
mtcars <- mtcars |> dplyr::group_by(cyl) |> dplyr::summarise(mpg = mean(mpg))
hgch_treemap(mtcars, var_cat = "cyl", var_num = "mpg")


}
\seealso{
\code{\link{data_draw}}
}
