% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CaCaNu}
\alias{hgch_treemap_CaCaNu}
\title{Treemap density by numeric variable}
\usage{
hgch_treemap_CaCaNu(data, title = NULL, subtitle = NULL,
  minColor = "#E63917", maxColor = "#18941E", reverse = TRUE,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Treemap density by numeric variable
}
\section{ftype}{
 Ca-Ca-Nu
}

\examples{
hgch_treemap_CaCaNu(sampleData("Ca-Ca-Nu",nrow = 10))
}
