% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_Cat}
\alias{hgch_treemap_Cat}
\title{Treemap (categories)}
\usage{
hgch_treemap_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  labelWrap = 12, colors = NULL, colorScale = "continuous",
  agg = "sum", marks = c(".", ","), nDigits = NULL, dropNa = FALSE,
  highlightValueColor = "#F9B233", percentage = FALSE, format = c("",
  ""), highlightValue = NULL, sliceN = NULL, showText = TRUE,
  showLegend = TRUE, legendPosition = c("right", "bottom"),
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, theme = NULL, lang = "es", ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_treemap_Cat(sampleData("Cat", nrow = 10))
}
