% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCatNum.R
\name{hgch_treemap_CatCat}
\alias{hgch_treemap_CatCat}
\title{treemap Chart Cat Cat Num}
\usage{
hgch_treemap_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
treemap Chart Cat Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_treemap_CatCat(sample_data("Cat-Cat", nrow = 10))
}
