% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatCatNum}
\alias{hgch_treemap_CatCatNum}
\title{Treemap (categories, categories, numbers)}
\usage{
hgch_treemap_CatCatNum(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num, Cat-Dat-Num,
}

\examples{
hgch_treemap_CatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
