% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatNum.R
\name{hgch_treemap_CatNum}
\alias{hgch_treemap_CatNum}
\title{Treemap Chart Cat Numeric}
\usage{
hgch_treemap_CatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Num, Yea-Num
}

\examples{
hgch_treemap_CatNum(sampleData("Cat-Num", nrow = 10))
}
