% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatNumP}
\alias{hgch_treemap_CatNumP}
\title{Treemap (categories, categories)}
\usage{
hgch_treemap_CatNumP(
  data,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  agg = "sum",
  colors = NULL,
  colorScale = "discrete",
  dropNaV = c(FALSE, FALSE),
  prefix = NULL,
  suffix = NULL,
  labelWrapV = c(12, 12),
  marks = c(".", ","),
  nDigits = NULL,
  percentage = FALSE,
  showText = TRUE,
  legend_show = TRUE,
  legendPosition = "center",
  theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE,
  lang = "es",
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-NumP, Yea-NumP, Dat-NumP
}

\examples{
hgch_treemap_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
