% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCatNum.R
\name{hgch_treemap_CatYea}
\alias{hgch_treemap_CatYea}
\title{treemap Chart Cat Yea}
\usage{
hgch_treemap_CatYea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
treemap Chart Cat Yea
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_treemap_CatYea(sample_data("Cat-Yea", nrow = 10))
}
