% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCatNum.R
\name{hgch_treemap_CatYeaNum}
\alias{hgch_treemap_CatYeaNum}
\title{treemap Chart Cat Yea Num}
\usage{
hgch_treemap_CatYeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
treemap Chart Cat Yea Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_treemap_CatYeaNum(sample_data("Cat-Yea-Num", nrow = 10))
}
