% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_treemap_cn}
\alias{hgch_treemap_cn}
\title{hgch_treemap_cn}
\usage{
hgch_treemap_cn(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  minColor = "#E63917", maxColor = "#18941E", reverse = TRUE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_treemap_cn
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_treemap_cn(sampleData("Ca-Nu",nrow = 10))
}

