% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_discrete_color_CaNu}
\alias{hgch_treemap_discrete_color_CaNu}
\title{hgch_treemap_discrete_color_CaNu}
\usage{
hgch_treemap_discrete_color_CaNu(data, title = "", export = FALSE)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_treemap_discrete_color_CaNu
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_treemap_discrete_color_CaNu(sampleData("Ca-Nu",nrow = 10))
}

