% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_discrete_color_Cat}
\alias{hgch_treemap_discrete_color_Cat}
\title{Treemap coloured by categorical variable}
\usage{
hgch_treemap_discrete_color_Cat(data, title = NULL, subtitle = NULL,
  caption = NULL, back_color = "white", color_title = "black",
  export = FALSE)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Treemap coloured by categorical variable
}
\section{ctypes}{

Cat
}

\examples{

hgch_treemap_discrete_color_Cat(sampleData("Cat", nrow = 10))

}
