% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_discrete_color_CatNum}
\alias{hgch_treemap_discrete_color_CatNum}
\title{Treemap coloured by first variable}
\usage{
hgch_treemap_discrete_color_CatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, back_color = "white", color_title = "black",
  export = FALSE)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Treemap coloured by first variable
}
\section{ctypes}{

Cat-Num
}

\examples{

hgch_treemap_discrete_color_CatNum(sampleData("Cat-Num", nrow = 10))

}
