% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_nested_CatCatNum}
\alias{hgch_treemap_nested_CatCatNum}
\title{Nested treemap density by numeric variable}
\usage{
hgch_treemap_nested_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, minColor = "#009EE3", maxColor = "#9B71AF",
  reverse = TRUE, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Nested treemap density by numeric variable
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{

hgch_treemap_nested_CatCatNum(sampleData("Cat-Cat-Num",nrow = 10))

}
