% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_nested_CatNumNum}
\alias{hgch_treemap_nested_CatNumNum}
\title{Nested treemap density}
\usage{
hgch_treemap_nested_CatNumNum(data, title = NULL, subtitle = NULL,
  caption = NULL, minColor = "#440154", maxColor = "#FDE725",
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Nested treemap density
}
\section{ctypes}{

Cat-Num-Num
}

\examples{

hgch_treemap_nested_CatNumNum(sampleData("Cat-Num-Num", nrow = 10))

}
