% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_nested_grouped_CatCatNum}
\alias{hgch_treemap_nested_grouped_CatCatNum}
\title{Nested treemap grouped by categorical variable}
\usage{
hgch_treemap_nested_grouped_CatCatNum(data, title = NULL,
  subtitle = NULL, caption = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Nested treemap grouped by categorical variable
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{

hgch_treemap_nested_grouped_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))

}
