% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-viz-opts.R
\name{hgch_treemap_options}
\alias{hgch_treemap_options}
\title{Treemap options}
\arguments{
\item{treemap_layout}{Option for algorithm to be used for setting position and dimensions of the points. Options: "stripes", "squarified", "strip", "sliceAndDice". Default is "squarified"}

\item{treemap_direction}{Defines which direction the layout algorithm will start drawing ("vertical" or "horizontal"). Default is vertical.}
}
\description{
This page describes the treemap options arguments
}
\keyword{internal}
