% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCat.R
\name{hgch_treemap_YeaCat}
\alias{hgch_treemap_YeaCat}
\title{Treemap chart Yea Cat}
\usage{
hgch_treemap_YeaCat(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read <\code{\link[=hgch_viz_options]{chart-options}}> a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_treemap_options]{treemap-options}}> which specifically contains the additional arguments
that work only for this type of chart}
}
\description{
\code{hgch_treemap_YeaCat()} Create a highcharter treemap plot based on a particular data type.
In this case, you can load data with only two columns, where the firts is a
\strong{year column} and the second must be a \strong{categorical column}, or be sure that firts two columns
they meet this condition, since it will be done a counting the categories of this columns.
}
\section{Ftype}{

Yea-Cat
}

\examples{
data <- sample_data("Yea-Cat", n = 30)
hgch_treemap_CatCat(data)

# Activate data labels
hgch_treemap_YeaCat(data,
                       dataLabels_show = TRUE)

# data with more of one column
data <- sample_data("Yea-Cat-Dat-Yea-Cat", n = 30)
hgch_treemap_YeaCat(data)

# Change variable to color and pallete type
hgch_treemap_YeaCat(data,
                       color_by = names(data)[2],
                       palette_type = "sequential")

# Change tooltip info and add additional information contained in your data
names_data <- names(data)
info_tool <- paste0("<b>",names_data[1],":</b> {", names_data[1],"}<br/><b>", names_data[4],":</b> {", names_data[4],"}<br/>")
data \%>\%
 hgch_treemap_YeaCat(tooltip = info_tool)

}
\concept{Yea-Cat plots}
