% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCatNum.R
\name{hgch_treemap_YeaCatNum}
\alias{hgch_treemap_YeaCatNum}
\title{treemap Chart Yea Cat Num}
\usage{
hgch_treemap_YeaCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
treemap Chart Yea Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_treemap_YeaCatNum(sample_data("Yea-Cat-Num", nrow = 10))
}
