% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatNum.R
\name{hgch_treemap_YeaNum}
\alias{hgch_treemap_YeaNum}
\title{treemap Chart Yea Num}
\usage{
hgch_treemap_YeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
treemap Chart Yea Num
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_treemap_YeaNum(sample_data("Yea-Num", nrow = 10))
}
