% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_waterfall_CaNu}
\alias{hgch_waterfall_CaNu}
\title{hgch_waterfall_CaNu}
\usage{
hgch_waterfall_CaNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_waterfall_CaNu
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_waterfall_CaNu(sampleData("Ca-Nu",nrow = 10))
}

