% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_waterfall_CatNum}
\alias{hgch_waterfall_CatNum}
\title{Waterfall}
\usage{
hgch_waterfall_CatNum(data, title = NULL, xAxisTitle = NULL,
  caption = NULL, aggregation = "sum", yAxisTitle = NULL,
  subtitle = NULL)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Waterfall
}
\section{ctypes}{

Cat-Num
}

\examples{
hgch_waterfall_CatNum(sampleData("Cat-Num", nrow = 10))
}
