% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_wordCloud_Cat.R
\name{hgch_wordCloud_Cat}
\alias{hgch_wordCloud_Cat}
\title{WordCloud chart Cat}
\usage{
hgch_wordCloud_Cat(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{<\code{\link[=hgch_viz_options]{chart-options}}> Options to configure your hgchmagic plots}
}
\description{
\code{hgch_pie_Cat()} Create a highcharter wordCloud based on a particular data type.
In this case, you can load data with only one \strong{categorical column} or \strong{text}
}
\section{Ftype}{

Cat
}

\seealso{
Other Cat plots: 
\code{\link{hgch_bar_Cat}()},
\code{\link{hgch_bubbles_Cat}()},
\code{\link{hgch_donut_Cat}()},
\code{\link{hgch_pie_Cat}()},
\code{\link{hgch_treemap_Cat}()}
}
\concept{Cat plots}
