% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgchc_treemap_CatCat}
\alias{hgchc_treemap_CatCat}
\title{Treemap (categories, categories)}
\usage{
hgchc_treemap_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, agg = "sum", colors = NULL,
  colorScale = "discrete", dropNaV = c(FALSE, FALSE), format = c("",
  ""), labelWrapV = c(12, 12), marks = c(".", ","), nDigits = NULL,
  percentage = FALSE, showText = TRUE, showLegend = TRUE,
  legendPosition = c("right", "bottom"), theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, lang = "es", ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat, Cat-Yea, Cat-Dat,
}

\examples{
hgch_treemap_CatCat(sampleData("Cat-Cat", nrow = 10))
}
