% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-prep.R
\docType{package}
\name{hgchmagic-package}
\alias{hgchmagic}
\alias{hgchmagic-package}
\title{hgchmagic: hgchmagic is a wrapper for hichcharter R package.}
\description{
With hgchmagic you can provide more info in form of list and improve your visualizations.
}
\keyword{internal}
