% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{hgchmagic_default_opts}
\alias{hgchmagic_default_opts}
\title{Chart Options}
\usage{
hgchmagic_default_opts(
  agg = "sum",
  agg_text = NULL,
  caption = NULL,
  palette_colors = NULL,
  color_by = NULL,
  drop_na = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  label_wrap = 12,
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  text_color = "#5A6B72",
  text_show = TRUE,
  text_size = 3,
  theme = NULL,
  title = NULL,
  ver_label = NULL,
  ver_line = NULL,
  opts = NULL
)
}
\arguments{
\item{agg}{defaults to "sum"}

\item{agg_text}{defaults to NULL}

\item{caption}{defaults to NULL}

\item{palette_colors}{defaults to NULL}

\item{color_by}{defaults to "discrete"}

\item{drop_na}{defaults to FALSE}

\item{highlight_value}{defaults to NULL}

\item{highlight_value_color}{defaults to '#F9B233'}

\item{hor_label}{defaults to NULL}

\item{hor_line}{defaults to NULL}

\item{label_wrap}{defaults to 12}

\item{marks}{defaults to c("." "")}

\item{n_digits}{defaults to NULL}

\item{order}{defaults to NULL}

\item{orientation}{defaults to "ver"}

\item{percentage}{defaults to FALSE}

\item{prefix}{defaults to NULL}

\item{slice_n}{defaults to NULL}

\item{sort}{defaults to "no"}

\item{subtitle}{defaults to NULL}

\item{suffix}{defaults to NULL}

\item{text_color}{defaults to "#5A6B72"}

\item{text_show}{defaults to TRUE}

\item{text_size}{defaults to 3}

\item{theme}{defaults to NULL}

\item{title}{defaults to NULL}

\item{ver_label}{defaults to NULL}

\item{ver_line}{defaults to NULL}

\item{opts}{defaults to NULL}

\item{format_num}{defaults to "1,500.00"}
}
\description{
Chart Options
}
