% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{hgchWhich}
\alias{hgchWhich}
\title{#' hgch_test_docs
#' test
#' @name hgch_test_docs
#' @export
#' @return The sum of \code{x} and \code{y}.
#' @section noftypes: Ca,Ca-Nu
#' @examples
#' add(1, 1)
hgch_test_docs <- NULL}
\usage{
hgchWhich(d)
}
\description{
#' hgch_test_docs2
#' test
#' @name hgch_test_docs2
#' @export
#' @return The sum of \code{x} and \code{y}.
#' @section ftypes: Ca,Ca-Ca-Ca
#' @examples
#' add(1, 1)
hgch_test_docs2 <- NULL
}
