% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-helpers.R
\name{list_bar}
\alias{list_bar}
\title{Extract Categorical and Numeric Data from a Data Frame}
\usage{
list_bar(data, frtype)
}
\arguments{
\item{data}{A data frame.}

\item{frType}{a string indicating the format of the data frame (e.g., "CatNum", "DatNum", "CatCatNum")}
}
\value{
A list containing a data frame that has the following columns: "name", "y", "..label", and "..color".
}
\description{
This function takes a data frame and extracts categorical and numeric data from four specified columns.
It returns a list of data that can be used for visualizations.
}
\keyword{internal}
