% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-opts.R
\name{plot_opts}
\alias{plot_opts}
\title{Prepare data for visualization}
\usage{
plot_opts(viz = NULL, ...)
}
\arguments{
\item{...}{additional arguments to be passed to \code{\link[dsvizopts]{dsviz_default_opts}}}
}
\value{
list with options to change properties of plot
}
\description{
This function prepares data for visualization by performing operations such as aggregation and data formatting.
}
\keyword{internal}
