% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_Cat}
\alias{hgch_area_Cat}
\title{Area (categories)}
\usage{
hgch_area_Cat(
  data = NULL,
  agg_text = NULL,
  allow_point = FALSE,
  background = "#ffffff",
  border_color = "#CCCCCC",
  border_width = 1,
  caption = NULL,
  click_function = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  cursor = NULL,
  drop_na = FALSE,
  export = FALSE,
  fill_opacity = 0.5,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = " ",
  label_wrap = 12,
  lang = "es",
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  spline = FALSE,
  start_zero = TRUE,
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = NULL,
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = " ",
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_area_Cat(sampleData("Cat", nrow = 10))
}
