% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_Cat}
\alias{hgch_area_Cat}
\title{Area (categories)}
\usage{
hgch_area_Cat(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_area_Cat(sampleData("Cat", nrow = 10))
}
