% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_CatCat}
\alias{hgch_area_CatCat}
\title{Area (categories, ordered categories, numbers)}
\usage{
hgch_area_CatCat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  orientation = "ver", graphType = "grouped", startAtZero = TRUE,
  agg = "sum", spline = FALSE, colors = NULL, colorOpacity = 0.5,
  dropNaV = c(FALSE, FALSE), format = c("", ""), labelWrapV = c(12,
  12), legendPosition = "right", marks = c(".", ","), nDigits = NULL,
  order1 = NULL, order2 = NULL, percentage = FALSE,
  showText = TRUE, theme = NULL, tooltip = list(headerFormat = NULL,
  pointFormat = NULL, shared = NULL), export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat
}

\examples{
hgch_area_CatCat(sampleData("Cat-Cat-Num", nrow = 10))
}
