% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_CatCatNum}
\alias{hgch_area_CatCatNum}
\title{Area (categories, ordered categories, numbers)}
\usage{
hgch_area_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  orientation = "ver", graphType = "grouped", startAtZero = TRUE,
  agg = "sum", spline = FALSE, colors = NULL, dropNa = c(FALSE,
  FALSE), format = c("", ""), labelWrap = c(12, 12),
  legendPosition = "right", marks = c(".", ","), nDigits = NULL,
  order1 = NULL, order2 = NULL, percentage = FALSE, theme = tma(),
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
hgch_area_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
