% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_CatDatNum.R
\name{hgch_area_CatDatNum}
\alias{hgch_area_CatDatNum}
\title{area Chart Cat Dat Numeric}
\usage{
hgch_area_CatDatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
area Chart Cat Dat Numeric
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
hgch_area_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
