% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_CatNum}
\alias{hgch_area_CatNum}
\title{Area (categories, numbers)}
\usage{
hgch_area_CatNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
hgch_area_CatNum(sampleData("Cat-Num", nrow = 10))
}
