% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_CatNum}
\alias{hgch_area_CatNum}
\title{Area (categories, numbers)}
\usage{
hgch_area_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  orientation = "ver", startAtZero = TRUE, labelWrap = 12,
  colors = NULL, colorOpacity = 0.5, agg = "sum", spline = FALSE,
  marks = c(".", ","), nDigits = NULL, dropNa = FALSE,
  percentage = FALSE, format = c("", ""), order = NULL,
  sort = "no", sliceN = NULL, showText = TRUE,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, lang = "es", theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
hgch_area_CatNum(sampleData("Cat-Num", nrow = 10))
}
