% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_CatNum}
\alias{hgch_area_CatNum}
\title{Vertical area}
\usage{
hgch_area_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  aggregation = "sum", startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical area
}
\section{ctypes}{

Cat-Num
}

\examples{

hgch_area_CatNum(sampleData("Cat-Num", nrow = 10))

}
