% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_CatYeaNum.R
\name{hgch_area_CatYeaNum}
\alias{hgch_area_CatYeaNum}
\title{area Chart Cat Cat Numeric}
\usage{
hgch_area_CatYeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num
}

\examples{
hgch_area_CatYeaNum(sample_data("Cat-Yea-Num", nrow = 100))
}
