% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_Dat.R
\name{hgch_area_Dat}
\alias{hgch_area_Dat}
\title{area Chart Dat Numeric}
\usage{
hgch_area_Dat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Dat
}

\examples{
hgch_area_Dat(sampleData("Dat", nrow = 10))
}
