% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_DatNum}
\alias{hgch_area_DatNum}
\title{Area (dates, numbers)}
\usage{
hgch_area_DatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  orientation = "ver", startAtZero = TRUE, labelWrap = 12,
  colors = NULL, colorOpacity = 0.5, agg = "sum", spline = FALSE,
  marks = c(".", ","), nDigits = NULL, dropNa = FALSE,
  percentage = FALSE, format = c("", ""), order = NULL,
  sort = "no", sliceN = NULL, tooltip = list(headerFormat = NULL,
  pointFormat = NULL), export = FALSE, theme = tma(), ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare a quantities over time (Year-month-day)
}
\section{ctypes}{

Dat-Num
}

\examples{
hgch_area_DatNum(sampleData("Dat-Num", nrow = 10))
}
