% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_stacked_100_CatCatNum}
\alias{hgch_area_stacked_100_CatCatNum}
\title{Vertical 100 stacked area}
\usage{
hgch_area_stacked_100_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  aggregation = "sum", startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical 100 stacked area
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{

hgch_area_stacked_100_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))

}
