% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{hgch_area_YeaNum}
\alias{hgch_area_YeaNum}
\title{Vertical area}
\usage{
hgch_area_YeaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  aggregation = "sum", startAtZero = FALSE, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical area
}
\section{ctypes}{

Yea-Num
}

\examples{

hgch_area_YeaNum(sampleData("Yea-Num", nrow = 10))

}
