% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_area_YeaNum.R
\name{hgch_area_YeaNum}
\alias{hgch_area_YeaNum}
\title{area Chart Yea Numeric}
\usage{
hgch_area_YeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Yea-Num, Yea-Num
}

\examples{
hgch_area_YeaNum(sample_data("Yea-Num", nrow = 10))
}
