% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{hgch_bar}
\alias{hgch_bar}
\title{Highcharter Bar Chart}
\usage{
hgch_bar(data, dic = NULL, var_cat = NULL, var_num = NULL, ...)
}
\arguments{
\item{data}{a data frame containing the variables to be plotted.}

\item{dic}{a named list, each element corresponding to a column name in \code{data} and defining a set of labels for that column.}

\item{var_cat}{a character vector with the names of categorical variables.}

\item{var_num}{a character vector with the names of numerical variables.}

\item{...}{additional arguments to be passed to \code{\link{plot_opts}}.}
}
\value{
a Highcharter bar chart.
}
\description{
Creates a bar chart using Highcharter library.
}
\examples{
data(mtcars)
mtcars$cyl <- as.character(mtcars$cyl)
mtcars <- mtcars |> group_by(cyl) |> summarize(mpg = mean(mpg))
hgch_bar(mtcars, var_cat = "cyl", var_num = "mpg")

}
