% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatNum.R
\name{hgch_bar_Cat}
\alias{hgch_bar_Cat}
\title{Bar Chart Cat}
\usage{
hgch_bar_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat
}

\examples{
hg_bar_Cat(sampleData("Cat", nrow = 10))
}
