% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_Cat}
\alias{hgch_bar_Cat}
\title{Bar (categories)}
\usage{
hgch_bar_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, horLineLabel = NULL,
  verLine = NULL, verLineLabel = NULL, marks = c(" ", "."), nDigits = 2,
  colors = c("#009EE3", "#F9B233"), dropNa = FALSE, format = "{value}",
  highlightValue = NULL, order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_bar_Cat(sampleData("Cat", nrow = 10))
}
