% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_Cat.R, R/hgch_bar_CatNum.R
\name{hgch_bar_Cat}
\alias{hgch_bar_Cat}
\title{Bar Chart Cat Numeric}
\usage{
hgch_bar_Cat(data, ...)

hgch_bar_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation

This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat, Yea-Num


Cat
}

\examples{
hgch_bar_Cat(sample_data("Cat", nrow = 10))
hgch_bar_Cat(sample_data("Cat", nrow = 10))
}
