% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{hgch_bar_Cat}
\alias{hgch_bar_Cat}
\title{Bar chart for a categorical variable}
\usage{
hgch_bar_Cat(data, ...)
}
\arguments{
\item{data}{a data frame.}

\item{...}{additional arguments to pass to the \code{hgch_bar} function.}
}
\value{
A highchart object.
}
\description{
This function creates a horizontal or vertical bar chart for a categorical variable.
}
\examples{
data(iris)
iris <- iris |> dplyr::select(Species)
hgch_bar_Cat(iris, percentage = TRUE)
}
