% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_Cat}
\alias{hgch_bar_Cat}
\title{Bar (categories)}
\usage{
hgch_bar_Cat(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_bar_Cat(sampleData("Cat", nrow = 10))
}
