% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatCat}
\alias{hgch_bar_CatCat}
\title{Grouped bar by first category (categories, categories)}
\usage{
hgch_bar_CatCat(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bar_CatCat(sampleData("Cat-Cat", nrow = 10))
}
