% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatCatNum.R
\name{hgch_bar_CatCat}
\alias{hgch_bar_CatCat}
\title{bar Chart Cat Cat Num}
\usage{
hgch_bar_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bar Chart Cat Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bar_CatCat(sample_data("Cat-Cat", nrow = 10))
}
