% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatCat.R
\name{hgch_bar_CatCat}
\alias{hgch_bar_CatCat}
\title{Bar Chart Cat Cat Numeric}
\usage{
hgch_bar_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat, Cat-Yea-Num
}

\examples{
hgch_bar_CatCat(sample_data("Cat-Cat", nrow = 10))
}
