% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatCat}
\alias{hgch_bar_CatCat}
\title{Grouped bar by first category (categories, categories)}
\usage{
hgch_bar_CatCat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  graphType = "grouped", agg = "sum", colors = c("#009EE3",
  "#F9B233"), colorScale = "discrete", dropNa = c(FALSE, FALSE),
  format = c("", ""), labelWrap = c(12, 12),
  legendPosition = "right", marks = c(".", ","), nDigits = NULL,
  order1 = NULL, order2 = NULL, orientation = "ver",
  percentage = FALSE, theme = tma(diffColorsBar = FALSE),
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bar_CatCat(sampleData("Cat-Cat", nrow = 10))
}
