% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatCatNum}
\alias{hgch_bar_CatCatNum}
\title{Grouped bar by first category (categories, ordered categories, numbers)}
\usage{
hgch_bar_CatCatNum(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
hgch_bar_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
