% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatCatNum.R
\name{hgch_bar_CatCatNum}
\alias{hgch_bar_CatCatNum}
\title{Bar Chart Cat Cat Numeric}
\usage{
hgch_bar_CatCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num
}

\examples{
hgch_bar_CatCatNum(sample_data("Cat-Cat-Num", nrow = 10))
}
