% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatCatNum}
\alias{hgch_bar_CatCatNum}
\title{Grouped bar by first category (categories, ordered categories, numbers)}
\usage{
hgch_bar_CatCatNum(
  data,
  agg = "sum",
  agg_text = NULL,
  allow_point = FALSE,
  background = "#ffffff",
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  drop_na_legend = FALSE,
  export = FALSE,
  graph_type = "grouped",
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = " ",
  label_wrap = 12,
  label_wrap_legend = 12,
  lang = "es",
  legend_position = "center",
  legend_show = TRUE,
  marks = c(".", ","),
  n_digits = NULL,
  null_color = "#f7f7f7",
  order1 = NULL,
  order2 = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = NULL,
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
hgch_bar_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
