% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar_CatNum.R
\name{hgch_bar_CatNum}
\alias{hgch_bar_CatNum}
\title{Bar Chart Cat Numeric}
\usage{
hgch_bar_CatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}

\item{...}{
  Arguments passed on to \code{\link[=hgchmagic_default_opts]{hgchmagic_default_opts}}
  \describe{
    \item{\code{agg}}{defaults to "sum"}
    \item{\code{agg_text}}{defaults to NULL}
    \item{\code{caption}}{defaults to NULL}
    \item{\code{palette_colors}}{defaults to NULL}
    \item{\code{color_by}}{defaults to "discrete"}
    \item{\code{drop_na}}{defaults to FALSE}
    \item{\code{highlight_value}}{defaults to NULL}
    \item{\code{highlight_value_color}}{defaults to '#F9B233'}
    \item{\code{hor_label}}{defaults to NULL}
    \item{\code{hor_line}}{defaults to NULL}
    \item{\code{label_wrap}}{defaults to 12}
    \item{\code{marks}}{defaults to c("." "")}
    \item{\code{n_digits}}{defaults to NULL}
    \item{\code{order}}{defaults to NULL}
    \item{\code{orientation}}{defaults to "ver"}
    \item{\code{percentage}}{defaults to FALSE}
    \item{\code{prefix}}{defaults to NULL}
    \item{\code{slice_n}}{defaults to NULL}
    \item{\code{sort}}{defaults to "no"}
    \item{\code{subtitle}}{defaults to NULL}
    \item{\code{suffix}}{defaults to NULL}
    \item{\code{text_color}}{defaults to "#5A6B72"}
    \item{\code{text_show}}{defaults to TRUE}
    \item{\code{text_size}}{defaults to 3}
    \item{\code{theme}}{defaults to NULL}
    \item{\code{title}}{defaults to NULL}
    \item{\code{ver_label}}{defaults to NULL}
    \item{\code{ver_line}}{defaults to NULL}
    \item{\code{opts}}{defaults to NULL}
  }}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Num, Yea-Num
}

\examples{
gg_bar_CatNum(sampleData("Cat-Num", nrow = 10))
}
