% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatNum}
\alias{hgch_bar_CatNum}
\title{Bar (categories, numbers)}
\usage{
hgch_bar_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, horLineLabel = NULL,
  verLine = NULL, verLineLabel = NULL, agg = "sum",
  colors = c("#009EE3", "#F9B233"), dropNa = FALSE, format = "{value}",
  highlightValue = NULL, order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
hgch_bar_CatNum(sampleData("Cat-Num", nrow = 10))
}
