% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatNum}
\alias{hgch_bar_CatNum}
\title{Bar (categories, numbers)}
\usage{
hgch_bar_CatNum(data = NULL, agg = "sum", agg_text = NULL,
  allow_point = FALSE, background = "#ffffff",
  border_color = "#CCCCCC", border_width = 1, caption = NULL,
  clickFunction = NULL, colors = NULL, color_click = NULL,
  color_hover = NULL, color_opacity = 0.7, color_scale = "discrete",
  cursor = NULL, drop_na = FALSE, export = FALSE,
  fill_opacity = 0.5, highlight_value = NULL,
  highlight_valueColor = "#F9B233", horLabel = NULL, horLine = NULL,
  horLine_label = " ", labelWrap = 12, lang = "es",
  legend_position = "center", legend_show = TRUE, marks = c(".",
  ","), nDigits = NULL, null_color = "#f7f7f7", order = NULL,
  orientation = "ver", percentage = FALSE, prefix = NULL,
  text_show = TRUE, sliceN = NULL, sort = "no", subtitle = NULL,
  suffix = NULL, title = NULL, theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  verLabel = NULL, verLine = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
hgch_bar_CatNum(sampleData("Cat-Num", nrow = 10))
}
