% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatNum}
\alias{hgch_bar_CatNum}
\title{Bar (categories, numbers)}
\usage{
hgch_bar_CatNum(
  data = NULL,
  agg = "sum",
  agg_text = NULL,
  allow_point = FALSE,
  background = "#ffffff",
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  export = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = " ",
  label_wrap = 12,
  lang = "es",
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = NULL,
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = " ",
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
hgch_bar_CatNum(sampleData("Cat-Num", nrow = 10))
}
