% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_CatNumP}
\alias{hgch_bar_CatNumP}
\title{Bar (ordered category, n numbers)}
\usage{
hgch_bar_CatNumP(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  graphType = "grouped", agg = "sum", colors = c("#009EE3",
  "#F9B233"), colorScale = "discrete", dropNaV = c(FALSE, FALSE),
  format = c("", ""), labelWrapV = c(12, 12), marks = c(".", ","),
  nDigits = NULL, order1 = NULL, order2 = NULL,
  orientation = "ver", percentage = FALSE,
  legendPosition = c("right", "bottom"), theme = tma(diffColorsBar =
  FALSE), tooltip = list(headerFormat = NULL, pointFormat = NULL, shared
  = NULL), export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Cat-NumP
}

\examples{
hgch_bar_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
