% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_DatNum}
\alias{hgch_bar_DatNum}
\title{Bar (dates, numbers)}
\usage{
hgch_bar_DatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, horLineLabel = NULL,
  verLine = NULL, verLineLabel = NULL, agg = "sum", orientation = "ver",
  marks = c("", "."), nDigits = NULL, dropNa = FALSE,
  colorHighlightValue = "#F9B233", percentage = FALSE, format = c("", ""),
  highlightValue = NULL, order = NULL, sort = "no", sliceN = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL), export = FALSE,
  thema = tma())
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities over dates
}
\section{ctypes}{

Dat-Num
}

\examples{
hgch_bar_DatNum(sampleData("Dat-Num", nrow = 10))
}
