% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_CatCat}
\alias{hgch_bar_grouped_CatCat}
\alias{hgch_bar_grouped_CatCat}
\title{Grouped bar (categories, ordered categories)}
\usage{
hgch_bar_grouped_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, dropNa = FALSE, order = NULL, orientation = "ver",
  sort = "no", sliceN = NULL, theme = NULL, export = FALSE, ...)

hgch_bar_grouped_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, dropNa = FALSE, order = NULL, orientation = "ver",
  sort = "no", sliceN = NULL, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}

\item{data}{A data.frame}
}
\value{
Highcharts visualization

Highcharts visualization
}
\description{
Compare quantities among two categories

Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat


Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bar_grouped_CatCat(sampleData("Cat-Cat", nrow = 10))
hgch_bar_grouped_CatCat(sampleData("Cat-Cat", nrow = 10))
}
