% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_CatCatNum}
\alias{hgch_bar_grouped_CatCatNum}
\title{Grouped bar by first category (categories, ordered categories, numbers)}
\usage{
hgch_bar_grouped_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  agg = "sum", colors = c("#009EE3", "#F9B233"),
  colorScale = "discrete", dropNa = c(FALSE, FALSE), format = c("",
  ""), labelWrap = c(12, 12), leyendPosition = "right",
  marks = c(".", ","), nDigits = NULL, order1 = NULL,
  order2 = NULL, orientation = "ver", percentage = FALSE,
  showText = TRUE, theme = tma(diffColorsBar = FALSE),
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
hgch_bar_grouped_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
