% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_CatDatNum}
\alias{hgch_bar_grouped_CatDatNum}
\title{Grouped bar (categories, dates, numbers)}
\usage{
hgch_bar_grouped_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, yLine = NULL,
  yLineLabel = NULL, agg = "sum", dropNa = FALSE, order = NULL,
  orientation = "ver", percentage = FALSE, sort = "no", sliceN = NULL,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among categories over dates
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
hgch_bar_grouped_CatDatNum(sampleData("Cat-Dat-Num", nrow = 10))
}
