% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_hor_CatNumP}
\alias{hgch_bar_grouped_hor_CatNumP}
\title{Horizontal grouped bar}
\usage{
hgch_bar_grouped_hor_CatNumP(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal grouped bar
}
\section{ftypes}{
 Cat-NumP
}
\examples{

hgch_bar_grouped_hor_CatNumP(sampleDatta("Cat-NumP",nrow = 10))

}

