% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_grouped_ver_CatCatNum}
\alias{hgch_bar_grouped_ver_CatCatNum}
\title{Vertical grouped bar}
\usage{
hgch_bar_grouped_ver_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical grouped bar
}
\section{ftype}{

Cat-Cat-Num
}
\examples{

hgch_bar_grouped_ver_CatCatNum(sampleDatta("Cat-Cat-Num",nrow = 10))

}

