% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_hor_CaNu}
\alias{hgch_bar_hor_CaNu}
\title{Horizontal bar}
\usage{
hgch_bar_hor_CaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "mean",
  back_color = "white", color_title = "black", theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal bar
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_bar_hor_CaNu(sampleData("Ca-Nu",nrow = 10))
}
