% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_hor_CatNum}
\alias{hgch_bar_hor_CatNum}
\title{Horizontal bar}
\usage{
hgch_bar_hor_CatNum(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = "mean", back_color = "white", color_title = "black",
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Horizontal bar
}
\section{ftype}{
 Cat-Num
}

\examples{
hgch_bar_hor_CatNum(sampleData("Cat-Num",nrow = 10))
}
