% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_NuP}
\alias{hgch_bar_NuP}
\alias{hgch_bar_ver_NuP}
\title{Vertical bar}
\usage{
hgch_bar_ver_NuP(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical bar
}
\section{ftypes}{
 NuP
}

\examples{
hgch_bar_NuP(sampleData("NuP",nrow = 10))
}
