% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_100_CatDatNum}
\alias{hgch_bar_stacked_100_CatDatNum}
\title{100% stacked bar (categories, dates, numbers)}
\usage{
hgch_bar_stacked_100_CatDatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = NULL, verLine = NULL, verLineLabel = NULL, agg = "sum",
  nDigits = 2, dropNa = FALSE, format = "{value}",
  legendLayout = "horizontal", order = NULL, orientation = "ver",
  sliceN = NULL, theme = NULL, tooltip = list(headerFormat = NULL,
  pointFormat = NULL, shared = NULL), export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among 100% stacked categories over dates
}
\section{ctypes}{

Cat-Dat-Num
}

\examples{
hgch_bar_stacked_100_CatYeaNum(sampleData("Cat-Dat-Num", nrow = 10))
}
