% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_CatCat}
\alias{hgch_bar_stacked_CatCat}
\title{Stacked bar by first cartegory (categories, categories)}
\usage{
hgch_bar_stacked_CatCat(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = NULL, verLine = NULL, verLineLabel = NULL,
  dropNa = FALSE, format = "{value}", legendLayout = "horizontal",
  order = NULL, orientation = "ver", percentage = FALSE,
  sort = "no", sliceN = NULL, theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare stacked categories
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bar_stacked_CatCat(sampleData("Cat-Cat", nrow = 10))
}
