% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_stacked_CatCatNum}
\alias{hgch_bar_stacked_CatCatNum}
\title{Stacked bar by first category (categories, ordered categories, numbers)}
\usage{
hgch_bar_stacked_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = NULL, verLine = NULL, verLineLabel = NULL,
  agg = "sum", dropNa = FALSE, format = "{value}",
  legendLayout = "horizontal", order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among stacked categories
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bar_stacked_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
