% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_CaNu}
\alias{hgch_bar_ver_CaNu}
\title{hgch_bar_ver_CaNu}
\usage{
hgch_bar_ver_CaNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, sort = "no", aggregate = NULL, startAtZero = FALSE,
  theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_ver_CaNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_bar_ver_CaNu(sampleData("Ca-Nu",nrow = 10))
}

