% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_CatNum}
\alias{hgch_bar_ver_CatNum}
\title{Vertical bar}
\usage{
hgch_bar_ver_CatNum(data, topn = NULL, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = NULL, back_color = "white", color_title = "black",
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Vertical bar
}
\section{ftype}{

Cat-Num
}

\examples{
hgch_bar_ver_CatNum(sampleDatta("Cat-Num",nrow = 10))
}
