% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_top_CaNu}
\alias{hgch_bar_ver_top_CaNu}
\title{Ordered vertical bar}
\usage{
hgch_bar_ver_top_CaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, reverse = TRUE,
  aggregate = "mean", theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Ordered vertical bar
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_bar_ver_top_CaNu(sampleData("Ca-Nu",nrow = 10))
}
