% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_ver_YeNu}
\alias{hgch_bar_ver_YeNu}
\title{hgch_bar_ver_YeNu}
\usage{
hgch_bar_ver_YeNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_bar_ver_YeNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_bar_ver_YeNu(sampleData("Ye-Nu",nrow = 10))
}

