% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bar_CatCatNum.R
\name{hgch_bar_YeaCat}
\alias{hgch_bar_YeaCat}
\title{bar Chart Yea Cat}
\usage{
hgch_bar_YeaCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bar Chart Yea Cat
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bar_YeaCat(sample_data("Yea-Cat", nrow = 10))
}
