% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_YeaNum}
\alias{hgch_bar_YeaNum}
\title{Bar (years, numbers)}
\usage{
hgch_bar_YeaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL, horLineLabel = NULL,
  verLine = NULL, verLineLabel = NULL, agg = "sum",
  colors = c("#009EE3", "#F9B233"), dropNa = FALSE, format = "{value}",
  highlightValue = NULL, order = NULL, orientation = "ver",
  percentage = FALSE, sort = "no", sliceN = NULL, theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities over years
}
\section{ctypes}{

Yea-Num
}

\examples{
hgch_bar_YeaNum(sampleData("Yea-Num", nrow = 10))
}
