% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{hgch_bar_YeaNum}
\alias{hgch_bar_YeaNum}
\title{Bar (years, numbers)}
\usage{
hgch_bar_YeaNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  labelWrap = 12, colors = NULL, colorScale = "no", agg = "sum",
  orientation = "ver", marks = c(".", ","), nDigits = NULL,
  dropNa = FALSE, highlightValueColor = "#F9B233",
  percentage = FALSE, format = c("", ""), highlightValue = NULL,
  order = NULL, sort = "no", sliceN = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, theme = tma(), ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities over years
}
\section{ctypes}{

Yea-Num
}

\examples{
hgch_bar_YeaNum(sampleData("Yea-Num", nrow = 10))
}
