% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_Cat.R
\name{hgch_bubbles_Cat}
\alias{hgch_bubbles_Cat}
\title{Bubbles Chart Cat Numeric}
\usage{
hgch_bubbles_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat
}

\examples{
hgch_bubbles_Cat(sample_data("Cat", nrow = 10))
}
