% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{hgch_bubbles_Cat}
\alias{hgch_bubbles_Cat}
\title{bubbles (categories)}
\usage{
hgch_bubbles_Cat(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_bubbles_Cat(sampleData("Cat", nrow = 10))
}
