% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{hgch_bubbles_CatCat}
\alias{hgch_bubbles_CatCat}
\title{Grouped bubbles by first category (categories, categories)}
\usage{
hgch_bubbles_CatCat(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Grouped bubbles by first category (categories, categories)
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bubbles_CatCat(sampleData("Cat-Cat", nrow = 10))
}
