% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbles.R
\name{hgch_bubbles_CatCat}
\alias{hgch_bubbles_CatCat}
\title{Grouped bubbles by first category (categories, categories)}
\usage{
hgch_bubbles_CatCat(
  data,
  agg_text = NULL,
  allow_point = TRUE,
  background = "#ffffff",
  bubble_min = 3,
  bubble_max = 10,
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_scale = "discrete",
  cursor = NULL,
  drop_na_v = c(FALSE, FALSE),
  export = FALSE,
  fill_opacity = 0.7,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  label_wrap_v = c(12, 12),
  lang = "es",
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Grouped bubbles by first category (categories, categories)
}
\section{ctypes}{

Cat-Cat, Cat-Dat, Cat-Yea, Yea-Cat, Yea-Dat, Yea-Yea, Dat-Cat, Dat-Yea, Dat-Dat
}

\examples{
hgch_bubbles_CatCat(sampleData("Cat-Cat", nrow = 10))
}
