% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatCatNum.R
\name{hgch_bubbles_CatCat}
\alias{hgch_bubbles_CatCat}
\title{bubbles Chart Cat Cat Num}
\usage{
hgch_bubbles_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bubbles Chart Cat Cat Num
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bubbles_CatCat(sample_data("Cat-Cat", nrow = 10))
}
