% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatCat.R
\name{hgch_bubbles_CatYea}
\alias{hgch_bubbles_CatYea}
\title{Bubbles chart Cat Yea}
\usage{
hgch_bubbles_CatYea(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read <\code{\link[=hgch_viz_options]{chart-options}}> a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_bubbles_options]{bubbles-options}}> which specifically contains the additional arguments
that work only for this type of chart}
}
\description{
\code{hgch_bubbles_CatYea()} Create a highcharter bubbles plot based on a particular data type.
In this case, you can load data with only two columns, where the firts is a
\strong{categorical column} and the second must be a \strong{year column}, or be sure that firts two columns
they meet this condition, since it will be done a counting the categories of this columns.
}
\section{Ftype}{

Cat-Yea
}

\examples{
data <- sample_data("Cat-Yea", n = 30)
hgch_bubbles_CatYea(data)

# Activate data labels
hgch_bubbles_CatYea(data,
                       dataLabels_show = TRUE)

# data with more of one column
data <- sample_data("Cat-Yea-Cat-Dat-Num-Cat", n = 30)
hgch_bubbles_CatYea(data)

# Change variable to color and pallete type
hgch_bubbles_CatYea(data,
                       color_by = names(data)[2],
                       palette_type = "sequential")

# Change tooltip info and add additional information contained in your data
names_data <- names(data)
info_tool <- paste0("<b>",names_data[1],":</b> {", names_data[1],"}<br/><b>", names_data[3],":</b> {", names_data[3],"}<br/>")
data \%>\%
 hgch_bubbles_CatYea(tooltip = info_tool)

}
\seealso{
Other Cat-Yea plots: 
\code{\link{hgch_area_CatYea}()},
\code{\link{hgch_bar_CatYea}()},
\code{\link{hgch_line_CatYea}()},
\code{\link{hgch_treemap_CatYea}()}
}
\concept{Cat-Yea plots}
