% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatNum.R
\name{hgch_bubbles_Yea}
\alias{hgch_bubbles_Yea}
\title{bubbles Chart Yea}
\usage{
hgch_bubbles_Yea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bubbles Chart Yea
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_bubbles_Yea(sample_data("Yea", nrow = 10))
}
