% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_bubbles_CatCatNum.R
\name{hgch_bubbles_YeaCat}
\alias{hgch_bubbles_YeaCat}
\title{bubbles Chart Yea Cat}
\usage{
hgch_bubbles_YeaCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
bubbles Chart Yea Cat
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_bubbles_YeaCat(sample_data("Yea-Cat", nrow = 10))
}
