% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_donut_Ca}
\alias{hgch_donut_Ca}
\title{hgch_donut_Ca}
\usage{
hgch_donut_Ca(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = "count", export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_donut_Ca
}
\section{ftype}{
 Ca
}
\examples{
hgch_donut_Ca(sampleData("Ca",nrow = 10))
}

