% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_donut_CaNu}
\alias{hgch_donut_CaNu}
\title{hgch_donut_CaNu}
\usage{
hgch_donut_CaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, back_color = "white",
  color_title = "black", leg_op = TRUE, sort = "no", aggregate = "sum",
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_donut_CaNu
}
\section{ftype}{
 Ca-Nu
}

\examples{
hgch_donut_CaNu(sampleData("Ca-Nu",nrow = 10))
}
