% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{hgch_donut_Cat}
\alias{hgch_donut_Cat}
\title{Donut (categories)}
\usage{
hgch_donut_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  dropNa = FALSE, order = NULL, sort = "no", sliceN = NULL,
  theme = NULL, export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing counts of categories
}
\section{ctypes}{

Cat
}

\examples{
hgch_donut_Cat(sampleData("Cat", nrow = 10))
}
