% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_donut_Cat.R
\name{hgch_donut_Cat}
\alias{hgch_donut_Cat}
\title{Donut chart Cat}
\usage{
hgch_donut_Cat(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read \code{\link[dsvizopts]{chart_viz_options}} a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_pie_options]{pie-options}}> which specifically contains the additional arguments
that work only for this type of chart.}
}
\description{
\code{hgch_donut_Cat()} Create a highcharter donut plot based on a particular data type.
In this case, you can load data with only two columns, where the firts column is a
\strong{categorical column} and the second must be  a \strong{numeric class column}, or be sure that
two firts columns they meet this condition
}
\section{Ftype}{

Cat
}

\examples{
data <- sample_data("Cat", n = 30)
hgch_donut_Cat(data)

example with iris data
data <- iris \%>\% select(Species, Petal.Width)
# this plot show the sum of petal width by species
hgch_donut_Cat(data, color_by = "Species")
# if you want to calculate the average instead of the sum,
# you can use "agg" param inside a function
hgch_donut_Cat(data, agg = "mean")

# data with more of one column
data <- sample_data("Cat-Dat-Cat-Cat", n = 30)
hgch_donut_Cat(data)

# calculate percentage
hgch_donut_Cat(data, percentage = TRUE)

# numeric format
hgch_donut_Cat(data, percentage = TRUE, format_sample_num = "1.234,")

# You can call the mean and percentage in the tooltip plot
num_name <- names(data)[2]
data \%>\%
hgch_donut_Cat(agg = "mean",
                tooltip = paste0("Average: {", num_name ,"} <br/> Percentage: {\%}\%"))
}
\seealso{
Other Cat plots: 
\code{\link{hgch_bar_Cat}()},
\code{\link{hgch_bubbles_Cat}()},
\code{\link{hgch_pie_Cat}()},
\code{\link{hgch_treemap_Cat}()},
\code{\link{hgch_wordCloud_Cat}()}
}
\concept{Cat plots}
