% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{hgch_donut_Cat}
\alias{hgch_donut_Cat}
\title{donut (categories)}
\usage{
hgch_donut_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  labelWrap = 12, colors = NULL, colorScale = "discrete",
  agg = "sum", marks = c(".", ","), nDigits = NULL, dropNa = FALSE,
  highlightValueColor = "#F9B233", percentage = FALSE, format = c("",
  ""), highlightValue = NULL, order = NULL, sort = "no",
  sliceN = NULL, showText = TRUE, legendPosition = c("right",
  "bottom"), tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, theme = tma(), ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_donut_Cat(sampleData("Cat", nrow = 10))
}
