% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_donut_Cat}
\alias{hgch_donut_Cat}
\title{hgch_donut_Cat}
\usage{
hgch_donut_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no",
  aggregate = "count", export = FALSE, theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_donut_Cat
}
\section{ftype}{
 Cat
}
\examples{
hgch_donut_Cat(sampleData("Cat",nrow = 10))
}

