% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_donut_CatNum.R
\name{hgch_donut_Cat}
\alias{hgch_donut_Cat}
\title{donut Chart Cat}
\usage{
hgch_donut_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
donut Chart Cat
}
\section{ctypes}{

Cat, Yea
}

\examples{
hgch_donut_Cat(sampleData("Cat", nrow = 10))
}
