% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{hgch_donut_CatNum}
\alias{hgch_donut_CatNum}
\title{donut (categories, numbers)}
\usage{
hgch_donut_CatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, labelWrap = 12, colors = NULL,
  colorScale = "discrete", agg = "sum", marks = c(".", ","),
  nDigits = NULL, dropNa = FALSE, highlightValueColor = "#F9B233",
  percentage = FALSE, prefix = NULL, suffix = NULL,
  highlightValue = NULL, order = NULL, sort = "no", sliceN = NULL,
  showText = FALSE, legendPosition = "center",
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, lang = "es", theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing quantities among categories
}
\section{ctypes}{

Cat-Num
}

\examples{
hgch_donut_CatNum(sampleData("Cat-Num", nrow = 10))
}
