% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_donut_CatNum}
\alias{hgch_donut_CatNum}
\title{hgch_donut_CatNum}
\usage{
hgch_donut_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "sum",
  export = FALSE, theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_donut_CatNum
}
\section{ftype}{

Cat
}
\examples{

hgch_donut_CatNum(sampleData("Cat-Num",nrow = 10))

}

