% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_dumbbell_CatNumNum.R
\name{hgch_dumbbell_CatCatNum}
\alias{hgch_dumbbell_CatCatNum}
\title{Dumbbell chart Cat Cat Num}
\usage{
hgch_dumbbell_CatCatNum(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read \code{\link[dsvizopts]{chart_viz_options}} a general options summary to configure your hgchmagic plots.}
}
\description{
\code{hgch_dumbbell_CatCatNum()} Create a highcharter dumbbell plot based on a particular data type.
In this case, you can load data with only three columns, where the firts and second columns are
\strong{categoricals columns} and the third must be  a \strong{numeric class column}, or be sure that
three firts columns they meet this condition
}
\section{Ftype}{

Cat-Cat-Num
}

\examples{
data <- sample_data("Cat-Cat-Num", n = 30)
hgch_dumbbell_CatCatNum(data)

# Activate data labels
hgch_dumbbell_CatCatNum(data,
                       dataLabels_show = TRUE)

# if you want to calculate the average instead of the sum, you can use agg inside a function
hgch_dumbbell_CatCatNum(data,
                       agg = "mean",
                       dataLabels_show = TRUE)

# data with more of one column
data <- sample_data("Cat-Cat-Num-Dat-Yea-Cat", n = 30)
hgch_dumbbell_CatCatNum(data)

# Change variable to color and pallete type
hgch_dumbbell_CatCatNum(data,
                       color_by = names(data)[2],
                       palette_type = "sequential")

# Change tooltip info and add additional information contained in your data
names_data <- names(data)
info_tool <- paste0("<b>",names_data[1],":</b> {", names_data[1],"}<br/><b>", names_data[4],":</b> {", names_data[4],"}<br/>")
data \%>\%
 hgch_dumbbell_CatCatNum(tooltip = info_tool)

}
\seealso{
Other Cat-Cat-Num plots: 
\code{\link{hgch_bar_CatCatNum}()},
\code{\link{hgch_bubbles_CatCatNum}()},
\code{\link{hgch_treemap_CatCatNum}()}
}
\concept{Cat-Cat-Num plots}
