% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{hgch_line}
\alias{hgch_line}
\title{Highcharter line Chart}
\usage{
hgch_line(data, dic = NULL, var_dat = NULL, var_num = NULL, ...)
}
\arguments{
\item{data}{a data frame containing the variables to be plotted.}

\item{dic}{a named list, each element corresponding to a column name in \code{data} and defining a set of labels for that column.}

\item{var_dat}{a character vector with the names of date variable.}

\item{var_num}{a character vector with the names of numerical variables.}

\item{...}{additional arguments to be passed to \code{\link[dsvizopts]{dsviz_default_opts}}}
}
\value{
a Highcharter line chart.
}
\description{
Creates a line chart using Highcharter library.
}
\seealso{
\code{\link{data_draw}}
}
