% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CaCaNu}
\alias{hgch_line_CaCaNu}
\title{hgch_line_CaCaNu}
\usage{
hgch_line_CaCaNu(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  symbol = NULL, startAtZero = FALSE, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_line_CaCaNu
}
\section{ftype}{
 Ca-Ca-Nu
}
\examples{
hgch_line_CaCaNu(sampleData("Ca-Ca-Nu",nrow = 10))
}

