% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CaDaNu}
\alias{hgch_line_CaDaNu}
\title{Line}
\usage{
hgch_line_CaDaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Line
}
\section{ftype}{
 Ca-Da-Nu
}

\examples{
hgch_line_CaDaNu(sampleData("Ca-Da-Nu",nrow = 10))
}
