% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CaDaNu}
\alias{hgch_line_CaDaNu}
\title{hgch_line_CaDaNu}
\usage{
hgch_line_CaDaNu(data, title = NULL, xAxisTitle = NULL, yAxisTitle = NULL,
  symbol = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_line_CaDaNu
}
\section{ftype}{
 Ca-Ye-Nu
}
\examples{
hgch_line_CaDaNu(sampleData("Ca-Da-Nu",nrow = 10))
}

