% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_Cat}
\alias{hgch_line_Cat}
\title{Lines (categories)}
\usage{
hgch_line_Cat(
  data,
  agg_text = NULL,
  allow_point = FALSE,
  background = "#ffffff",
  caption = NULL,
  click_function = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  cursor = NULL,
  drop_na = FALSE,
  export = FALSE,
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = " ",
  label_wrap = 12,
  lang = "es",
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  spline = FALSE,
  start_zero = TRUE,
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = NULL,
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = " ",
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_line_Cat(sampleData("Cat", nrow = 10))
}
