% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_Cat}
\alias{hgch_line_Cat}
\title{Lines (categories)}
\usage{
hgch_line_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  orientation = "ver", startAtZero = TRUE, labelWrap = 12,
  colors = NULL, agg = "sum", spline = FALSE, marks = c(".", ","),
  nDigits = NULL, dropNa = FALSE, percentage = FALSE,
  format = c("", ""), order = NULL, sort = "no", sliceN = NULL,
  showText = TRUE, tooltip = list(headerFormat = NULL, pointFormat =
  NULL), export = FALSE, lang = "es", theme = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_line_Cat(sampleData("Cat", nrow = 10))
}
