% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatCat}
\alias{hgch_line_CatCat}
\title{Line (categories, ordered categories, numbers)}
\usage{
hgch_line_CatCat(data, opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat
}

\examples{
hgch_line_CatCat(sampleData("Cat-Cat-Num", nrow = 10))
}
