% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatCat}
\alias{hgch_line_CatCat}
\title{Line (categories, ordered categories, numbers)}
\usage{
hgch_line_CatCat(data, agg_text = NULL, allow_point = FALSE,
  background = "#ffffff", border_color = "#CCCCCC", border_width = 1,
  caption = NULL, clickFunction = NULL, colors = NULL,
  color_click = NULL, color_hover = NULL, color_opacity = 0.7,
  color_scale = "discrete", cursor = NULL, drop_naV = c(FALSE,
  FALSE), export = FALSE, fill_opacity = 0.5, highlight_value = NULL,
  highlight_valueColor = "#F9B233", horLabel = NULL, horLine = NULL,
  horLine_label = " ", labelWrapV = c(12, 12), lang = "es",
  legend_position = "center", legend_show = TRUE, marks = c(".",
  ","), nDigits = NULL, null_color = "#f7f7f7", order1 = NULL,
  order2 = NULL, orientation = "ver", percentage = FALSE,
  prefix = NULL, text_show = TRUE, sliceN = NULL, sort = "no",
  spline = FALSE, startAtZero = TRUE, subtitle = NULL,
  suffix = NULL, title = NULL, theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  verLabel = NULL, verLine = NULL, verLine_label = " ",
  opts = NULL, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat
}

\examples{
hgch_line_CatCat(sampleData("Cat-Cat-Num", nrow = 10))
}
