% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatCatNum}
\alias{hgch_line_CatCatNum}
\title{Line (categories, ordered categories, numbers)}
\usage{
hgch_line_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  orientation = "ver", startAtZero = TRUE, agg = "sum",
  spline = FALSE, colors = NULL, dropNaV = c(FALSE, FALSE),
  format = c("", ""), labelWrapV = c(12, 12),
  legendPosition = "right", marks = c(".", ","), nDigits = NULL,
  order1 = NULL, order2 = NULL, percentage = FALSE,
  showText = TRUE, theme = NULL, tooltip = list(headerFormat = NULL,
  pointFormat = NULL, shared = NULL), export = FALSE, lang = "es", ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among two categories
}
\section{ctypes}{

Cat-Cat-Num, Cat-Dat-Num, Cat-Yea-Num, Yea-Cat-Num, Yea-Dat-Num, Yea-Yea-Num, Dat-Cat-Num, Dat-Yea-Num, Dat-Dat-Num
}

\examples{
hgch_line_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
