% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_CatDat.R
\name{hgch_line_CatDat}
\alias{hgch_line_CatDat}
\title{line Chart Cat Dat Numeric}
\usage{
hgch_line_CatDat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
line Chart Cat Dat Numeric
}
\section{ctypes}{

Cat-Dat
}

\examples{
hgch_line_CatDat(sampleData("Cat-Dat", nrow = 10))
}
