% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatNum}
\alias{hgch_line_CatNum}
\title{Line}
\usage{
hgch_line_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Line
}
\section{ftype}{

Cat-Num
}

\examples{

hgch_line_CatNum(sampleDatta("Cat-Num",nrow = 10))

}
