% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatNumP}
\alias{hgch_line_CatNumP}
\title{Line (ordered category, n numbers)}
\usage{
hgch_line_CatNumP(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  orientation = "ver", startAtZero = TRUE, agg = "sum",
  spline = FALSE, colors = NULL, dropNaV = c(FALSE, FALSE),
  format = c("", ""), labelWrapV = c(12, 12),
  legendPosition = "right", marks = c(".", ","), nDigits = NULL,
  order1 = NULL, order2 = NULL, percentage = FALSE,
  showText = TRUE, theme = NULL, tooltip = list(headerFormat = NULL,
  pointFormat = NULL, shared = NULL), export = FALSE, lang = "es", ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare n quantities among category's levels
}
\section{ctypes}{

Cat-NumP
}

\examples{
hgch_line_CatNumP(sampleData("Cat-NumP", nrow = 10))
}
