% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_CatYeaNum}
\alias{hgch_line_CatYeaNum}
\title{Line (categories, years, numbers)}
\usage{
hgch_line_CatYeaNum(data, title = NULL, subtitle = NULL,
  caption = NULL, horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  startAtZero = TRUE, agg = "sum", spline = FALSE, colors = NULL,
  dropNa = c(FALSE, FALSE), format = c("", ""), labelWrap = c(12,
  12), leyendPosition = "right", marks = c(".", ","), nDigits = NULL,
  order1 = NULL, order2 = NULL, percentage = FALSE,
  showText = TRUE, theme = tma(), plotBandsFromX = NULL,
  plotBandsToX = NULL, plotBandsColorX = "rgba(68, 170, 213, .2)",
  plotBandsFromY = NULL, plotBandsToY = NULL,
  plotBandsColorY = "rgba(68, 170, 213, .2)",
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare quantities among categories over years
}
\section{ctypes}{

Cat-Yea-Num
}

\examples{
hgch_line_CatYeaNum(sampleData("Cat-Yea-Num", nrow = 10))
}
