% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_DatNum.R
\name{hgch_line_Dat}
\alias{hgch_line_Dat}
\title{line Chart Dat Numeric}
\usage{
hgch_line_Dat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Dat
}

\examples{
hgch_line_Dat(sampleData("Dat", nrow = 10))
}
