% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_DatNum.R
\name{hgch_line_DatNum}
\alias{hgch_line_DatNum}
\title{line Chart Dat Numeric}
\usage{
hgch_line_DatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Num, Yea-Num
}

\examples{
hgch_line_CatNum(sampleData("Dat-Num", nrow = 10))
}
