% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_DatNum}
\alias{hgch_line_DatNum}
\title{Line}
\usage{
hgch_line_DatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Line
}
\section{ctypes}{

Dat-Num
}

\examples{

hgch_line_DatNum(sampleData("Dat-Num", nrow = 10))

}
