% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_line_DatNum}
\alias{hgch_line_DatNum}
\title{Line (dates, numbers)}
\usage{
hgch_line_DatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  horLabel = NULL, verLabel = NULL, horLine = NULL,
  horLineLabel = " ", verLine = NULL, verLineLabel = " ",
  startAtZero = TRUE, labelWrap = 12, colors = NULL, agg = "sum",
  spline = FALSE, marks = c(".", ","), nDigits = NULL,
  dropNa = FALSE, percentage = FALSE, format = c("", ""),
  order = NULL, sort = "no", sliceN = NULL, plotBandsFromX = NULL,
  plotBandsToX = NULL, plotBandsColorX = "rgba(68, 170, 213, .2)",
  plotBandsFromY = NULL, plotBandsToY = NULL,
  plotBandsColorY = "rgba(68, 170, 213, .2)",
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, theme = tma(), ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare a quantities over time (Year-month-day)
}
\section{ctypes}{

Dat-Num
}

\examples{
hgch_line_DatNum(sampleData("Dat-Num", nrow = 10))
}
