% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_YeaNum.R
\name{hgch_line_Yea}
\alias{hgch_line_Yea}
\title{line Chart Yea}
\usage{
hgch_line_Yea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Yea
}

\examples{
hgch_line_Yea(sample_data("Yea", nrow = 10))
}
