% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_line_YeaNum.R
\name{hgch_line_YeaNum}
\alias{hgch_line_YeaNum}
\title{line Chart Yea Numeric}
\usage{
hgch_line_YeaNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Yea-Num, Yea-Num
}

\examples{
hgch_line_YeaNum(sample_data("Yea-Num", nrow = 10))
}
