% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_multilines.}
\alias{hgch_2yline_YeNuNu}
\alias{hgch_multilines.}
\alias{hgch_multilines_YeNuP}
\title{hgch_2yline_YeNuNu
2 y lines}
\usage{
hgch_2yline_YeNuNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle1 = NULL, yAxisTitle2 = NULL, symbol = NULL, export = FALSE,
  ...)

hgch_multilines_YeNuP(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL, symbol = NULL, startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}

\item{x}{A data.frame}
}
\value{
highcharts viz

highcharts viz
}
\description{
hgch_2yline_YeNuNu
2 y lines

hgch_multilines_YeNuP
Multilines
}
\section{ftypes}{
 Ye-Nu-Nu

 Ye-Nu*
}
\examples{
hgch_2yline_YeNuNu(sampleData("Ye-Nu-Nu",nrow = 10))
hgch_multilines_YeNuP(sampleData("Ye-Nu-Nu",nrow = 10))
}

