% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_multilines.}
\alias{hgch_multilines.}
\alias{hgch_multilines_YeNuP}
\title{hgch_multilines_YeNuP
Multilines}
\usage{
hgch_multilines_YeNuP(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_multilines_YeNuP
Multilines
}
\section{ftypes}{
 Ye-NuP
}

\examples{
hgch_multilines_YeNuP(sampleData("Ye-Nu-Nu",nrow = 10))
}
