% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{hgch_multilines_YeaNumP}
\alias{hgch_multilines_YeaNumP}
\title{Grouped line}
\usage{
hgch_multilines_YeaNumP(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, symbol = NULL,
  startAtZero = FALSE, theme = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Grouped line
}
\section{ctypess}{
 Yea-NumP
}

\examples{

hgch_multilines_YeaNumP(sampleDatta("Yea-NumP",nrow = 10))

}
