% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_pie_CaNu}
\alias{hgch_pie_CaNu}
\title{hgch_pie_CaNu}
\usage{
hgch_pie_CaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "sum",
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_pie_CaNu
}
\section{ftype}{
 Ca
}
\examples{
hgch_pie_CaNu(sampleData("Ca",nrow = 10))
}

