% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_pie_CatNum.R
\name{hgch_pie_Cat}
\alias{hgch_pie_Cat}
\title{Pie Chart Cat}
\usage{
hgch_pie_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat, Yea
}

\examples{
hgch_pie_Cat(sampleData("Cat", nrow = 10))
}
