% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_pie_Cat.R
\name{hgch_pie_Cat}
\alias{hgch_pie_Cat}
\title{Pie Chart Cat Numeric}
\usage{
hgch_pie_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat, Yea-Num
}

\examples{
hgch_pie_Cat(sample_data("Cat", nrow = 10))
}
