% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{hgch_pie_Cat}
\alias{hgch_pie_Cat}
\title{pie (categories)}
\usage{
hgch_pie_Cat(data, opts = opts, ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare category's levels
}
\section{ctypes}{

Cat, Yea, Dat
}

\examples{
hgch_pie_Cat(sampleData("Cat", nrow = 10))
}
