% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_pie_Cat}
\alias{hgch_pie_Cat}
\title{hgch_pie_Cat}
\usage{
hgch_pie_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  font_size = "13px", sort = "no", aggregate = "count", export = FALSE,
  theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_pie_Cat
}
\section{ctypes}{

Cat
}

\examples{

hgch_pie_Cat(sampleData("Cat",nrow = 10))

}
