% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_pie_CatNum}
\alias{hgch_pie_CatNum}
\title{hgch_pie_CatNum}
\usage{
hgch_pie_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, sort = "no", aggregate = "sum",
  export = FALSE, font_size = "13px", theme = NULL, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_pie_CatNum
}
\section{ctypes}{

Cat-Num
}

\examples{

hgch_pie_CatNum(sampleData("Cat-Num",nrow = 10))

}
