% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{hgch_pie_CatNum}
\alias{hgch_pie_CatNum}
\title{Pie (categories, numbers)}
\usage{
hgch_pie_CatNum(
  data,
  agg = "sum",
  agg_text = NULL,
  background = "#ffffff",
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  export = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  label_wrap = 12,
  lang = "es",
  legend_position = "center",
  legend_show = TRUE,
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing quantities among categories
}
\section{ctypes}{

Cat-Num
}

\examples{
hgch_pie_CatNum(sampleData("Cat-Num", nrow = 10))
}
