% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{hgch_pie_CatNum}
\alias{hgch_pie_CatNum}
\title{Pie (categories, numbers)}
\usage{
hgch_pie_CatNum(data, title = NULL, subtitle = NULL, caption = NULL,
  labelWrap = 12, colors = NULL, colorScale = "discrete",
  agg = "sum", marks = c(".", ","), nDigits = NULL, dropNa = FALSE,
  highlightValueColor = "#F9B233", percentage = FALSE, format = c("",
  ""), highlightValue = NULL, order = NULL, sort = "no",
  sliceN = NULL, legendPosition = c("right", "bottom"),
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, theme = tma(), ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Comparing quantities among categories
}
\section{ctypes}{

Cat-Num
}

\examples{
hgch_pie_CatNum(sampleData("Cat-Num", nrow = 10))
}
