% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_pie_Cat.R
\name{hgch_pie_Yea}
\alias{hgch_pie_Yea}
\title{pie Chart Yea}
\usage{
hgch_pie_Yea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
pie Chart Yea
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_pie_Yea(sample_data("Yea", nrow = 10))
}
