% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_polarcolumn_Ca}
\alias{hgch_polarcolumn_Ca}
\title{hgch_polarcolumn_Ca}
\usage{
hgch_polarcolumn_Ca(data, title = "")
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_polarcolumn_Ca
}
\section{ftype}{
 Ca
}

\examples{
hgch_polarcolumn_Ca(sampleData("Ca",nrow = 10))
}
