% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_sankey_CatCat.R
\name{hgch_sankey_CatCat}
\alias{hgch_sankey_CatCat}
\title{Sankey Cat Cat}
\usage{
hgch_sankey_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Sankey Cat Cat
}
\section{}{
NA
}

\examples{
hgch_sankey_CatCat(sample_data("Cat-Cat"))
}
