% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_scatter.R
\name{hgch_scatter}
\alias{hgch_scatter}
\title{Highcharter scatter Chart}
\usage{
hgch_scatter(
  data,
  dic = NULL,
  var_cat = NULL,
  var_dat = NULL,
  var_num = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame containing the variables to be plotted.}

\item{dic}{a named list, each element corresponding to a column name in \code{data} and defining a set of labels for that column.}

\item{var_cat}{the name of the categorical variable in the data frame}

\item{var_dat}{a character vector with the names of date variable.}

\item{var_num}{a character vector with the names of numerical variables.}

\item{...}{additional arguments to be passed to \code{\link[dsvizopts]{dsviz_default_opts}}}
}
\value{
a Highcharter scatter chart.
}
\description{
Creates a scatter chart using Highcharter library.
}
\examples{

data <- ggplot2::diamonds
data <- data |>
  dplyr::group_by(clarity) |>
  dplyr::summarise(x = sum(x, na.rm = T),
                   y = sum(carat, na.rm = T))

hgch_scatter(data,
             var_num = c("x", "y"),
             var_cat = "clarity")

 # Custom theme
  test_theme <- list(
    theme = list(
      palette_colors = c("#ffa92a", "lightgreen"),
      subtitle_align = "center",
      subtitle_family = "Roboto",
      subtitle_size = 15,
      subtitle_color = "#3b83b8",
      subtitle_weight = 700
    )
  )

  hgch_scatter(data,
               var_num = c("x", "y"),
               var_cat = "clarity",
               opts = test_theme,
               title = "test title",
               subtitle = "test subtitle")

}
\seealso{
\code{\link{data_draw}}
}
