% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CaCaNuNu}
\alias{hgch_scatter_CaCaNuNu}
\title{hgch_scatter_CaCaNuNu}
\usage{
hgch_scatter_CaCaNuNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, theme = NULL, export = FALSE,
  ...)
}
\description{
hgch_scatter_CaCaNuNu
}
\section{ftype}{
 Ca-Nu-Nu
}

