% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CaCaNuNuNu}
\alias{hgch_scatter_CaCaNuNuNu}
\title{hgch_scatter_CaCaNuNuNu}
\usage{
hgch_scatter_CaCaNuNuNu(data, title = NULL, subtitle = NULL,
  caption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, theme = NULL,
  export = FALSE, ...)
}
\description{
hgch_scatter_CaCaNuNuNu
}
\section{ftype}{
 Ca-Nu-Nu
}

