% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{hgch_scatter_CaNuNu}
\alias{hgch_scatter_CaNuNu}
\title{hgch_scatter_CaNuNu}
\usage{
hgch_scatter_CaNuNu(data, title = NULL, xAxisTitle = NULL,
  yAxisTitle = NULL)
}
\description{
hgch_scatter_CaNuNu
}
\section{ftype}{
 Ca-Nu-Nu
}

