% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CaNuNuNu}
\alias{hgch_scatter_CaNuNuNu}
\title{hgch_scatter_CaNuNuNu}
\usage{
hgch_scatter_CaNuNuNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, theme = NULL, export = FALSE,
  ...)
}
\description{
hgch_scatter_CaNuNuNu
}
\section{ftype}{
 Ca-Nu-Nu-Nu
}

