% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CatCatNumNum}
\alias{hgch_scatter_CatCatNumNum}
\title{Scatter grouped}
\usage{
hgch_scatter_CatCatNumNum(data, title = NULL, subtitle = NULL,
  Catption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Scatter grouped
}
\section{ctypes}{

Cat-Cat-Num-Num
}

\examples{
hgch_scatter_CatCatNumNum(sampleData("Cat-Cat-Num-Num",nrow = 10))
}
