% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_scatter_CatDatNum.R
\name{hgch_scatter_CatDat}
\alias{hgch_scatter_CatDat}
\title{scatter Chart Cat Dat}
\usage{
hgch_scatter_CatDat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
scatter Chart Cat Dat
}
\section{ctypes}{

Cat-Dat
}

\examples{
hgch_scatter_CatDat(sampleData("Cat-Dat", nrow = 10))
}
