% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_scatter_CatDatNum.R
\name{hgch_scatter_CatDatNum}
\alias{hgch_scatter_CatDatNum}
\title{Scatter chart Cat Dat Num}
\usage{
hgch_scatter_CatDatNum(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read <\code{\link[=hgch_viz_options]{chart-options}}> a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_scatter_options]{scatter-options}}> which specifically contains the additional arguments
that work only for this type of chart}
}
\description{
\code{hgch_scatter_CatDatNum()} Create a highcharter scatter plot based on a particular data type.
In this case, you can load data with only three columns,  where the firts is a \strong{categorical column},
second is a \strong{date column}, and the three is a \strong{numeric column}, or be sure that three firts columns
they meet this condition.
}
\section{Ftype}{

Dat-Num
}

\examples{
data <- sample_data("Cat-Dat-Num", n = 30)
hgch_scatter_CatDatNum(data)

# data with more of one column
data <- sample_data("Cat-Dat-Num-Dat-Cat-Cat", n = 30)
hgch_scatter_CatDatNum(data)

}
\seealso{
Other Cat-Dat-Num plots: 
\code{\link{hgch_area_CatDatNum}()},
\code{\link{hgch_line_CatDatNum}()}
}
\concept{Cat-Dat-Num plots}
