% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CatNumNum}
\alias{hgch_scatter_CatNumNum}
\title{Scatter}
\usage{
hgch_scatter_CatNumNum(data, title = NULL, subtitle = NULL,
  Catption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Scatter
}
\section{ctypes}{

Cat-Num-Num
}

\examples{

hgch_scatter_CatNumNum(sampleData("Cat-Num-Num",nrow = 10))

}
