% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CatNumNum}
\alias{hgch_scatter_CatNumNum}
\title{Scatter Cat Num Num}
\usage{
hgch_scatter_CatNumNum(
  data = NULL,
  agg = "sum",
  agg_text = NULL,
  allow_point = FALSE,
  background = "#ffffff",
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_opacity = 0.7,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  export = FALSE,
  fill_opacity = 0.5,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  hor_label = NULL,
  hor_line = NULL,
  hor_line_label = " ",
  label_wrap = 12,
  lang = "es",
  legend_position = "center",
  legend_show = TRUE,
  marks = c(".", ","),
  n_digits = NULL,
  n_digits_size = NULL,
  n_digits_y = NULL,
  n_digits_x = NULL,
  orientation = "ver",
  percentage = FALSE,
  prefix = NULL,
  prefix_size = NULL,
  prefix_x = NULL,
  prefix_y = NULL,
  regression = FALSE,
  regression_color = "#d35400",
  regression_equation = TRUE,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  spline = FALSE,
  subtitle = NULL,
  suffix = NULL,
  suffix_x = NULL,
  suffix_y = NULL,
  suffix_size = NULL,
  title = NULL,
  theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  ver_label = NULL,
  ver_line = NULL,
  ver_line_label = " ",
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{}
}
\description{
Scatter Cat Num Num
}
