% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CatNumNum}
\alias{hgch_scatter_CatNumNum}
\title{Scatter Cat Num Num}
\usage{
hgch_scatter_CatNumNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{}
}
\description{
Scatter Cat Num Num
}
