% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_scatter_CatNumNum.R
\name{hgch_scatter_CatNumNum}
\alias{hgch_scatter_CatNumNum}
\title{Scatter chart Num Num}
\usage{
hgch_scatter_CatNumNum(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read \code{\link[dsvizopts]{chart_viz_options}} a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_scatter_options]{scatter-options}}> which specifically contains the additional arguments
that work only for this type of chart.}
}
\description{
\code{hgch_scatter_CatNumNum()} Create a highcharter scatter plot based on a particular data type.
In this case, you can load data with only three columns,  where the first is
a \strong{categorical column}, and second and third columns are
\strong{numeric class}, or be sure that two firts columns they meet this condition
}
\section{Ftype}{

Cat-Num-Num
}

\examples{
data <- sample_data("Cat-Num-Num", n = 30)
hgch_scatter_CatNumNum(data)

example with iris data
data <- iris \%>\% select(Species, Petal.Width, Petal.Length)
hgch_scatter_CatNumNum(data)

# data with more of one column
data <- sample_data("Cat-Num-Num-Dat-Cat-Cat", n = 30)
hgch_scatter_CatNumNum(data)

}
\seealso{
Other Cat-Num-Num plots: 
\code{\link{hgch_dumbbell_CatNumNum}()}
}
\concept{Cat-Num-Num plots}
