% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CatNumNumNum}
\alias{hgch_scatter_CatNumNumNum}
\title{Scatter Cat Num Num Num}
\usage{
hgch_scatter_CatNumNumNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{}
}
\description{
Scatter Cat Num Num Num
}
