% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_CatNumNumNum}
\alias{hgch_scatter_CatNumNumNum}
\title{SCattter}
\usage{
hgch_scatter_CatNumNumNum(data, title = NULL, subtitle = NULL,
  Catption = NULL, xAxisTitle = NULL, yAxisTitle = NULL, theme = NULL,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
SCattter
}
\section{ftype}{
 Cat-Num-Num-Num
}

\examples{
hgch_scatter_CatNumNumNum(sampleData("Cat-Num-Num-Num",nrow = 10))
}
