% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_scatter_NumNum.R
\name{hgch_scatter_NumNum}
\alias{hgch_scatter_NumNum}
\title{Scatter Chart Cat Cat Numeric}
\usage{
hgch_scatter_NumNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num
}

\examples{
hgch_scatter_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
