% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{hgch_scatter_NumNumNum}
\alias{hgch_scatter_NumNumNum}
\title{Scatter Num Num Num}
\usage{
hgch_scatter_NumNumNum(data = NULL, opts = NULL, ...)
}
\arguments{
\item{data}{}
}
\description{
Scatter Num Num Num
}
