% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CaCaNu}
\alias{hgch_treemap_CaCaNu}
\title{hgch_treemap_CaCaNu}
\usage{
hgch_treemap_CaCaNu(data, title = NULL, minColor = "#E63917",
  maxColor = "#18941E", reverse = TRUE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_treemap_CaCaNu
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_treemap_CaCaNu(sampleData("Ca-Nu",nrow = 10))
}

