% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CaNu}
\alias{hgch_treemap_CaNu}
\title{hgch_treemap_CaNu}
\usage{
hgch_treemap_CaNu(data, title = NULL, subtitle = NULL, caption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, minColor = "#E63917",
  maxColor = "#18941E", reverse = TRUE, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
hgch_treemap_CaNu
}
\section{ftype}{
 Ca-Nu
}
\examples{
hgch_treemap_CaNu(sampleData("Ca-Nu",nrow = 10))
}

