% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatNum.R
\name{hgch_treemap_Cat}
\alias{hgch_treemap_Cat}
\title{Treemap Chart Cat}
\usage{
hgch_treemap_Cat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
Treemap Chart Cat
}
\section{ctypes}{

Cat, Yea
}

\examples{
hgch_treemap_Cat(sampleData("Cat", nrow = 10))
}
