% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_Cat}
\alias{hgch_treemap_Cat}
\title{Treemap density coloured by categorical variable}
\usage{
hgch_treemap_Cat(data, title = NULL, subtitle = NULL, caption = NULL,
  minColor = "#E63917", maxColor = "#18941E", back_color = "white",
  color_title = "black", reverse = TRUE, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Treemap density coloured by categorical variable
}
\section{ctypes}{

Cat
}

\examples{

hgch_treemap_Cat(sampleData("Cat", nrow = 10))

}
