% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_treemap.R
\name{hgch_treemap_Cat}
\alias{hgch_treemap_Cat}
\title{treemap chart for a categorical variable}
\usage{
hgch_treemap_Cat(data, ...)
}
\arguments{
\item{data}{a data frame.}

\item{...}{additional arguments to pass to the \code{hgch_treemap} function.}
}
\value{
A highchart object.
}
\description{
This function creates a horizontal or vertical treemap chart for a categorical variable.
}
\examples{
data(iris)
iris <- iris |> dplyr::select(Species)
hgch_treemap_Cat(iris, percentage = TRUE)
}
