% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatCat}
\alias{hgch_treemap_CatCat}
\title{Treemap (categories, categories)}
\usage{
hgch_treemap_CatCat(
  data,
  agg_text = NULL,
  allow_point = FALSE,
  background = "#ffffff",
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  drop_na_legend = FALSE,
  export = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  label_wrap = 12,
  label_wrap_legend = 12,
  lang = "es",
  legend_position = "center",
  legend_show = TRUE,
  marks = c(".", ","),
  n_digits = NULL,
  order1 = NULL,
  order2 = NULL,
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat, Cat-Yea, Cat-Dat,
}

\examples{
hgch_treemap_CatCat(sampleData("Cat-Cat", nrow = 10))
}
