% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCat.R
\name{hgch_treemap_CatCat}
\alias{hgch_treemap_CatCat}
\title{Treemap Chart Cat Cat Numeric}
\usage{
hgch_treemap_CatCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat, Cat-Yea-Num
}

\examples{
hgch_treemap_CatCat(sampleData("Cat-Cat", nrow = 10))
}
