% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatCatNum}
\alias{hgch_treemap_CatCatNum}
\title{Treemap (categories, categories, numbers)}
\usage{
hgch_treemap_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, agg = "sum", colors = NULL,
  colorScale = "discrete", dropNaV = c(FALSE, FALSE), prefix = NULL,
  suffix = NULL, labelWrapV = c(12, 12), marks = c(".", ","),
  nDigits = NULL, percentage = FALSE, showText = TRUE,
  showLegend = TRUE, legendPosition = "center", theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL, shared = NULL),
  export = FALSE, lang = "es", ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num, Cat-Dat-Num,
}

\examples{
hgch_treemap_CatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
