% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatCatNum}
\alias{hgch_treemap_CatCatNum}
\title{Treemap (categories, categories, numbers)}
\usage{
hgch_treemap_CatCatNum(
  data,
  agg = "sum",
  agg_text = NULL,
  background = "#ffffff",
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  drop_na_v = c(FALSE, FALSE),
  export = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  label_wrap_v = c(12, 12),
  lang = "es",
  legend_position = "center",
  legend_show = TRUE,
  marks = c(".", ","),
  n_digits = NULL,
  order1 = NULL,
  order2 = NULL,
  percentage = FALSE,
  prefix = NULL,
  text_show = TRUE,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  title = NULL,
  theme = NULL,
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num, Cat-Dat-Num,
}

\examples{
hgch_treemap_CatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
