% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatCatNum}
\alias{hgch_treemap_CatCatNum}
\title{Treemap density by numeric variable}
\usage{
hgch_treemap_CatCatNum(data, title = NULL, subtitle = NULL,
  minColor = "#E63917", maxColor = "#18941E", reverse = TRUE,
  export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Treemap density by numeric variable
}
\section{ftype}{

Cat-Cat-Num
}
\examples{

hgch_treemap_CatCatNum(sampleData("Cat-Cat-Num",nrow = 10))

}

