% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatCatNum}
\alias{hgch_treemap_CatCatNum}
\title{Treemap coloured by first variable and grouped by second variable}
\usage{
hgch_treemap_CatCatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, agg = "sum", percentage = FALSE, theme = NULL,
  colors = NULL, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Treemap coloured by first variable and grouped by second variable
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{

hgch_treemap_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))

}
