% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCatNum.R
\name{hgch_treemap_CatCatNum}
\alias{hgch_treemap_CatCatNum}
\title{Treemap Chart Cat Cat Numeric}
\usage{
hgch_treemap_CatCatNum(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
This chart does not allow for chaning orientation
}
\section{ctypes}{

Cat-Cat-Num, Cat-Yea-Num
}

\examples{
hgch_treemap_CatCatNum(sampleData("Cat-Cat-Num", nrow = 10))
}
