% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatNum}
\alias{hgch_treemap_CatNum}
\title{Treemap (categories, numbers)}
\usage{
hgch_treemap_CatNum(data, title = NULL, subtitle = NULL,
  caption = NULL, labelWrap = 12, colors = NULL,
  colorScale = "continuous", agg = "sum", marks = c(".", ","),
  nDigits = NULL, dropNa = FALSE, highlightValueColor = "#F9B233",
  percentage = FALSE, prefix = NULL, suffix = NULL,
  highlightValue = NULL, sliceN = NULL, showText = TRUE,
  showLegend = TRUE, legendPosition = c("right", "bottom"),
  tooltip = list(headerFormat = NULL, pointFormat = NULL),
  export = FALSE, theme = NULL, lang = "es", ...)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
hgch_treemap_CatNum(sampleData("Cat-Num", nrow = 10))
}
