% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatNum}
\alias{hgch_treemap_CatNum}
\title{Treemap density by numeric variable}
\usage{
hgch_treemap_CatNum(data, title = NULL, subtitle = NULL, Catption = NULL,
  xAxisTitle = NULL, yAxisTitle = NULL, minColor = "#E63917",
  maxColor = "#18941E", back_color = "white", color_title = "black",
  reverse = TRUE, export = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}
}
\value{
highcharts viz
}
\description{
Treemap density by numeric variable
}
\section{ftype}{

Cat-Num
}

\examples{

hgch_treemap_CatNum(sampleData("Cat-Num",nrow = 10))

}
