% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{hgch_treemap_CatNum}
\alias{hgch_treemap_CatNum}
\title{Treemap (categories, numbers)}
\usage{
hgch_treemap_CatNum(
  data,
  agg = "sum",
  agg_text = NULL,
  background = "#ffffff",
  border_color = "#CCCCCC",
  border_width = 1,
  caption = NULL,
  clickFunction = NULL,
  colors = NULL,
  color_click = NULL,
  color_hover = NULL,
  color_scale = "discrete",
  cursor = NULL,
  drop_na = FALSE,
  export = FALSE,
  highlight_value = NULL,
  highlight_value_color = "#F9B233",
  label_wrap = 12,
  lang = "es",
  marks = c(".", ","),
  n_digits = NULL,
  order = NULL,
  percentage = FALSE,
  prefix = NULL,
  slice_n = NULL,
  sort = "no",
  subtitle = NULL,
  suffix = NULL,
  text_show = TRUE,
  title = NULL,
  theme = NULL,
  tooltip = NULL,
  opts = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame}
}
\value{
Highcharts visualization
}
\description{
Compare aggregations among category's levels
}
\section{ctypes}{

Cat-Num, Yea-Num, Dat-Num,
}

\examples{
hgch_treemap_CatNum(sampleData("Cat-Num", nrow = 10))
}
