% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCatNum.R
\name{hgch_treemap_CatYeaNum}
\alias{hgch_treemap_CatYeaNum}
\title{Treemap chart Cat Yea Num}
\usage{
hgch_treemap_CatYeaNum(data, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a
lazy data frame (e.g. from dbplyr or dtplyr), or fringe data (e.g from homodatum).}

\item{...}{Read <\code{\link[=hgch_viz_options]{chart-options}}> a general options summary to configure your hgchmagic plots
and <\code{\link[=hgch_treemap_options]{treemap-options}}> which specifically contains the additional arguments
that work only for this type of chart}
}
\description{
\code{hgch_treemap_CatYeaNum()} Create a highcharter treemap plot based on a particular data type.
In this case, you can load data with only three columns, where the firts column is a
\strong{categorical column}, second is a \strong{year column} and the third must be  a \strong{numeric class column},
or be sure that three firts columns they meet this condition
}
\section{Ftype}{

Cat-Yea-Num
}

\examples{
data <- sample_data("Cat-Yea-Num", n = 30)
hgch_treemap_CatYeaNum(data)

# Activate data labels
hgch_treemap_CatYeaNum(data,
                       dataLabels_show = TRUE)

# data with more of one column
data <- sample_data("Cat-Yea-Num-Dat-Yea-Cat", n = 30)
hgch_treemap_CatYeaNum(data)

# Change variable to color and pallete type
hgch_treemap_CatYeaNum(data,
                       color_by = names(data)[2],
                       palette_type = "sequential")

# Change tooltip info and add additional information contained in your data
names_data <- names(data)
info_tool <- paste0("<b>",names_data[1],":</b> {", names_data[1],"}<br/><b>", names_data[4],":</b> {", names_data[4],"}<br/>")
data \%>\%
 hgch_treemap_CatYeaNum(tooltip = info_tool)

}
