% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatNum.R
\name{hgch_treemap_Yea}
\alias{hgch_treemap_Yea}
\title{treemap Chart Yea}
\usage{
hgch_treemap_Yea(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
treemap Chart Yea
}
\section{ctypes}{

Yea, Yea
}

\examples{
hgch_treemap_Yea(sample_data("Yea", nrow = 10))
}
