% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch_treemap_CatCatNum.R
\name{hgch_treemap_YeaCat}
\alias{hgch_treemap_YeaCat}
\title{treemap Chart Yea Cat}
\usage{
hgch_treemap_YeaCat(data, ...)
}
\arguments{
\item{data}{A data.frame}
}
\description{
treemap Chart Yea Cat
}
\section{ctypes}{

Cat-Cat-Num
}

\examples{
hgch_treemap_YeaCat(sample_data("Yea-Cat", nrow = 10))
}
