% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-prep.R
\docType{package}
\name{hgchmagic-package}
\alias{hgchmagic}
\alias{hgchmagic-package}
\title{hgchmagic: A wrapper for highcharter}
\description{
A wrapper for highcharter
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/datasketch/hgchmagic/issues}
}

}
\keyword{internal}
